### 2021-05-21 #3303 @whedon
[PRE REVIEW]: spopt: a python package for solving spatial optimization problems in PySAL
**Submitting author:** @xf37 (<a href="http://orcid.org/0000-0001-7253-3154">Xin Feng</a>)
**Repository:** <a href="https://github.com/pysal/spopt" target ="_blank">https://github.com/pysal/spopt</a>
**Version:** v0.1.1
