### 2021-05-21 #3303 @kyleniemeyer
Hi @xf37, I think the issue with your paper is in the YAML block at the top—there is an indentation error with the `orcid` field for Sergio Rey.
