### 2021-06-02 #3330 @whedon
[REVIEW]: spopt: a python package for solving spatial optimization problems in PySAL
**Submitting author:** @xf37 (<a href="http://orcid.org/0000-0001-7253-3154">Xin Feng</a>)
**Repository:** https://github.com/pysal/spopt
**Branch with paper.md** (empty if default branch):
**Version:** 0.4.1
