### 2021-06-14 3814d70
chore: init arch of locate package
