### 2021-06-17 1276a3f
chore: test coverage models using facloc.ipynb
