+++ spopt/locate/__init__.py
@@ -1,0 +1,1 @@
+from .coverage import LSCP, MCLP
