### 2021-06-17 pysal/spopt#164 @jGaboardi
remove draft_pdf action
The action should now be removed because the `spopt` JOSS paper is [under review](https://github.com/openjournals/joss-reviews/issues/3330) and should be generated by `whedon` in the review issue.
