### 2021-06-17 b3709af
update & add JOSS reference DOIs
