+++ paper/paper.bib
@@ -6,21 +6,22 @@
   number={7},
   pages={797--811},
   year={2006},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/13658810600665111}
 }
 
 @article{duque2012max,
-  author = {Duque, Juan C. and Anselin, Luc and Rey, Sergio J.},
-  title = {THE MAX-P-REGIONS PROBLEM*},
-  journal = {Journal of Regional Science},
-  volume = {52},
-  number = {3},
-  pages = {397--419},
-  doi = {https://doi.org/10.1111/j.1467-9787.2011.00743.x},
-  url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-9787.2011.00743.x},
-  eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-9787.2011.00743.x},
-  abstract = {ABSTRACT In this paper, we introduce a new spatially constrained clustering problem called the max-p-regions problem. It involves the clustering of a set of geographic areas into the maximum number of homogeneous regions such that the value of a spatially extensive regional attribute is above a predefined threshold value. We formulate the max-p-regions problem as a mixed integer programming (MIP) problem, and propose a heuristic solution.},
-  year = {2012}
+  author={Duque, Juan C. and Anselin, Luc and Rey, Sergio J.},
+  title={THE MAX-P-REGIONS PROBLEM*},
+  journal={Journal of Regional Science},
+  volume={52},
+  number={3},
+  pages={397--419},
+  doi={10.1111/j.1467-9787.2011.00743.x},
+  url={https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-9787.2011.00743.x},
+  eprint={https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-9787.2011.00743.x},
+  abstract={ABSTRACT In this paper, we introduce a new spatially constrained clustering problem called the max-p-regions problem. It involves the clustering of a set of geographic areas into the maximum number of homogeneous regions such that the value of a spatially extensive regional attribute is above a predefined threshold value. We formulate the max-p-regions problem as a mixed integer programming (MIP) problem, and propose a heuristic solution.},
+  year={2012}
 }
 
 @article{koopmans1949optimum,
@@ -29,7 +30,8 @@
   journal={Econometrica: Journal of the Econometric Society},
   pages={136--146},
   year={1949},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.2307/1907301}
 }
 
 @article{miller1960integer,
@@ -40,7 +42,8 @@
   number={4},
   pages={326--329},
   year={1960},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/321043.321046}
 }
 
 @article{murray2019contemporary,
@@ -49,7 +52,8 @@
   journal={Omega},
   pages={102176},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.omega.2019.102176}
 }
 
 @article{wei2020efficient,
@@ -58,7 +62,8 @@
   journal={International Journal of Geographical Information Science},
   pages={1--17},
   year={2020},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/13658816.2020.1759806}
 }
 
 @article{openshaw1977geographical,
@@ -67,7 +72,8 @@
   journal={Transactions of the Institute of British Geographers},
   pages={459--472},
   year={1977},
[…]
