+++ paper/paper.bib
@@ -117,3 +117,11 @@
   doi={10.31219/osf.io/yzt2p}
 }
 
+@article{Rey2021,
+  author={Rey, Sergio J. and Anselin, Luc and Amaral, Pedro and Arribas-Bel, Dani and Cortes, Renan Xavier and Gaboardi, James David and Kang, Wei and Knaap, Elijah and Li, Ziqi and Lumnitz, Stefanie and Oshan, Taylor M. and Shao, Hu and Wolf, Levi John},
+  title={{The PySAL Ecosystem: Philosophy and Implementation}},
+  journal={Geographical Analysis},
+  year={2021},
+  doi={10.1111/gean.12276},
+  abstract={PySAL is a library for geocomputation and spatial data science. Written in Python, the library has a long history of supporting novel scholarship and broadening methodological impacts far afield of academic work. Recently, many new techniques, methods of analyses, and development modes have been implemented, making the library much larger and more encompassing than that previously discussed in the literature. As such, we provide an introduction to the library as it stands now, as well as the scientific and conceptual underpinnings of its core set of components. Finally, we provide a prospective look at the library's future evolution.}
+}
