### 2021-06-17 #3330 @jGaboardi
> Hi @xf37 it seems that several DOIs are missing from your submission and one is invalid as you can see above. Please fix this issue.

We have fixed the issue with missing DOIs.
