### 2021-06-17 7ce53a1
use micromambda for CI
