+++ .github/workflows/unittests.yml
@@ -22,34 +22,28 @@
             os: windows-latest
           - environment-file: .ci/38.yaml
             os: ubuntu-latest
-     defaults:
-       run:
-         shell: bash -l {0}
+    
      steps:
-       - uses: actions/checkout@v2
-       - uses: actions/cache@v2
-         env:
-           CACHE_NUMBER: 0
+       - name: checkout repo
+         uses: actions/checkout@v2
+       
+       - name: setup micromamba
+         uses: mamba-org/provision-with-micromamba@main
          with:
-           path: ~/conda_pkgs_dir
-           key: ${{ matrix.os }}-conda-${{ env.CACHE_NUMBER }}-${{ hashFiles(matrix.environment-file) }}
-       - uses: conda-incubator/setup-miniconda@v2
-         with:
-            miniconda-version: 'latest'
-            mamba-version: '*'
-            channels: conda-forge
-            channel-priority: true
-            auto-update-conda: false
-            auto-activate-base: false
-            environment-file: ${{ matrix.environment-file }}
-            activate-environment: test
-            use-only-tar-bz2: true
-       - run: mamba info --all
-       - run: mamba list
-       - run: conda config --show-sources
-       - run: conda config --show
-       - run: pytest -v spopt --cov=spopt --cov-report=xml
-       - name: codecov (${{ matrix.os }}, ${{ matrix.environment-file }})
+           environment-file: ${{ matrix.environment-file }}
+           micromamba-version: 'latest'
+    
+      - name: run tests - bash
+         shell: bash -l {0}
+         run: pytest -v spopt --cov=spopt --cov-report=xml
+         if: matrix.os != 'windows-latest'
+       
+       - name: run tests - powershell
+         shell: powershell
+         run: pytest -v spopt --cov=spopt --cov-report=xml
+         if: matrix.os == 'windows-latest'
+       
+       - name: codecov
          uses: codecov/codecov-action@v1
          with:
            token: ${{ secrets.CODECOV_TOKEN }}
