### 2021-06-17 c3703d4
tidy up CI envs
