+++ .ci/38.yaml
@@ -13,13 +13,8 @@ dependencies:
   - scikit-learn>=0.22
   - scipy>=1.0
   - spenc
-  # testing, docs, etc.
+  # testing
   - codecov
   - coverage
-  - nbsphinx
-  - numpydoc
   - pytest
   - pytest-cov
-  - sphinx>=1.4.3
-  - sphinxcontrib-bibtex==1.0.0
-  - sphinx_bootstrap_theme
