### 2021-06-17 1f7ecd9
add 39-dev env (libpysal dev)
