### 2021-06-17 4d4014d
add libpysal dev test
