+++ .github/workflows/unittests.yml
@@ -32,7 +32,12 @@
          with:
            environment-file: ${{ matrix.environment-file }}
            micromamba-version: 'latest'
-    
+       
+       - name: install bleeding edge libpysal (only Ubuntu / Python 3.9)
+         shell: bash -l {0}
+         run: pip install git+https://github.com/pysal/libpysal.git@master
+         if: matrix.os == 'ubuntu-latest' && contains(matrix.environment-file, 'DEV')
+       
        - name: run tests - bash
          shell: bash -l {0}
          run: pytest -v spopt --cov=spopt --cov-report=xml
