+++ .github/workflows/unittests.yml
@@ -18,10 +18,13 @@
          os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
          environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
          exclude:
-          - environment-file: .ci/39.yaml
-            os: windows-latest
-          - environment-file: .ci/38.yaml
-            os: ubuntu-latest
+           - environment-file: .ci/39.yaml
+             os: windows-latest
+           - environment-file: .ci/38.yaml
+             os: ubuntu-latest
+         include:
+           - environment-file: ci/39-DEV.yaml
+             os: ubuntu-latest
     
      steps:
        - name: checkout repo
