+++ .pre-commit-config.yaml
@@ -3,4 +3,4 @@ repos:
     rev: stable
     hooks: 
     - id: black
-      language_version: python3.8
+      language_version: python3.9
