+++ spopt/tests/test_azp.py
@@ -3,6 +3,7 @@ import libpysal
 import geopandas as gpd
 from spopt.region import AZP
 
+
 def test_azp():
     pth = libpysal.examples.get_path("mexicojoin.shp")
     mexico = gpd.read_file(pth)
@@ -13,5 +14,39 @@ def test_azp():
     model = AZP(mexico, w, attrs_name, n_clusters=5)
     model.solve()
     labels = np.array(
-        [0, 0, 2, 1, 1, 2, 2, 2, 3, 3, 3, 1, 2, 4, 4, 2, 4, 3, 3, 3, 4, 3, 0, 0, 2, 2, 2, 1, 2, 1, 1, 4])
+        [
+            0,
+            0,
+            2,
+            1,
+            1,
+            2,
+            2,
+            2,
+            3,
+            3,
+            3,
+            1,
+            2,
+            4,
+            4,
+            2,
+            4,
+            3,
+            3,
+            3,
+            4,
+            3,
+            0,
+            0,
+            2,
+            2,
+            2,
+            1,
+            2,
+            1,
+            1,
+            4,
+        ]
+    )
     return np.array_equal(model.labels_, labels)
