+++ spopt/tests/test_spenc.py
@@ -6,17 +6,16 @@ import unittest
 
 
 class TestSpenc(unittest.TestCase):
-
-
     def setUp(self):
         self.pth = libpysal.examples.get_path("mexicojoin.shp")
         self.mexico = gpd.read_file(self.pth)
         self.mexico["count"] = 1
         self.attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
         self.w = libpysal.weights.Queen.from_dataframe(self.mexico)
-        self.model = Spenc(gdf=self.mexico, w=self.w, attrs_name=self.attrs_name, random_state=123456)
+        self.model = Spenc(
+            gdf=self.mexico, w=self.w, attrs_name=self.attrs_name, random_state=123456
+        )
         self.model.solve()
-
 
     def test_labels_(self):
         observed_labels = self.model.labels_
