### 2021-06-18 8d88160
fix: classmethod best_practices in coverage.py
Revision made by ljwof in 18/06
