+++ spopt/tests/test_maxp.py
@@ -3,56 +3,69 @@ import libpysal
 import geopandas as gpd
 import os
 import sys
+import unittest
 
-# sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), ".")))
 from spopt.region import MaxPHeuristic
 
 
-def test_MaxPHeuristic():
+# Mexican states
+pth = libpysal.examples.get_path("mexicojoin.shp")
+MEXICO = gpd.read_file(pth)
 
-    pth = libpysal.examples.get_path("mexicojoin.shp")
-    mexico = gpd.read_file(pth)
-    mexico["count"] = 1
-    w = libpysal.weights.Queen.from_dataframe(mexico)
-    attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
-    threshold = 4
-    top_n = 2
-    threshold_name = "count"
-    numpy.random.seed(123456)
-    model = MaxPHeuristic(mexico, w, attrs_name, threshold_name, threshold, top_n)
-    model.solve()
-    labels = [
-        5,
-        5,
-        3,
-        6,
-        6,
-        6,
-        1,
-        1,
-        1,
-        1,
-        8,
-        6,
-        8,
-        2,
-        2,
-        8,
-        2,
-        8,
-        7,
-        7,
-        2,
-        7,
-        5,
-        3,
-        3,
-        5,
-        3,
-        4,
-        4,
-        4,
-        4,
-        7,
-    ]
-    numpy.array_equal(model.labels_, labels)
+
+class TestMaxPHeuristic(unittest.TestCase):
+    def setUp(self):
+
+        self.mexico = MEXICO.copy()
+
+        # labels for non-verbose and verbose:
+        # count=1 , threshold=4, top_n=2
+        self.basic_labels = [5, 5, 3, 6, 6, 6, 1, 1, 1, 1, 8, 6, 8, 2, 2, 8]
+        self.basic_labels += [2, 8, 7, 7, 2, 7, 5, 3, 3, 5, 3, 4, 4, 4, 4, 7]
+
+        # labels for:
+        # count=4 , threshold=10, top_n=5
+        self.complex_labels = [3, 3, 2, 10, 10, 2, 6, 9, 1, 6, 6, 10, 7, 4, 4, 7]
+        self.complex_labels += [4, 7, 1, 1, 9, 1, 3, 8, 8, 2, 8, 2, 5, 5, 5, 9]
+
+    def test_maxp_heuristic_basic_nonverbose(self):
+        self.mexico["count"] = 1
+        w = libpysal.weights.Queen.from_dataframe(self.mexico)
+        attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
+        threshold = 4
+        top_n = 2
+        threshold_name = "count"
+        numpy.random.seed(123456)
+        args = (self.mexico, w, attrs_name, threshold_name, threshold, top_n)
+        model = MaxPHeuristic(*args)
+        model.solve()
+
+        numpy.array_equal(model.labels_, self.basic_labels)
+
+    def test_maxp_heuristic_basic_verbose(self):
+        self.mexico["count"] = 1
+        w = libpysal.weights.Queen.from_dataframe(self.mexico)
+        attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
+        threshold = 4
+        top_n = 2
+        threshold_name = "count"
+        numpy.random.seed(123456)
+        args = (self.mexico, w, attrs_name, threshold_name, threshold, top_n)
+        model = MaxPHeuristic(*args, verbose=True)
+        model.solve()
+
+        numpy.array_equal(model.labels_, self.basic_labels)
+
+    def test_maxp_heuristic_complex(self):
[…]
