### 2021-06-19 6ca768b
blacken and clean AZP docs
