+++ spopt/region/azp.py
@@ -11,105 +11,121 @@ import networkx as nx
 
 from spopt.region.csgraph_utils import sub_adj_matrix, neighbors, is_connected
 from spopt.region.objective_function import ObjectiveFunctionPairwise
-from spopt.region.azp_util import AllowMoveStrategy, \
-                                            AllowMoveAZP,\
-                                            AllowMoveAZPSimulatedAnnealing
-from spopt.region.util import array_from_df_col, array_from_dict_values, \
-    assert_feasible, boolean_assert_feasible, copy_func, count, generate_initial_sol, \
-    make_move, Move, pop_randomly_from, random_element_from,\
-    scipy_sparse_matrix_from_w, separate_components, w_from_gdf,\
-    array_from_graph_or_dict, scipy_sparse_matrix_from_dict
+from spopt.region.azp_util import (
+    AllowMoveStrategy,
+    AllowMoveAZP,
+    AllowMoveAZPSimulatedAnnealing,
+)
+from spopt.region.util import (
+    array_from_df_col,
+    array_from_dict_values,
+    assert_feasible,
+    boolean_assert_feasible,
+    copy_func,
+    count,
+    generate_initial_sol,
+    make_move,
+    Move,
+    pop_randomly_from,
+    random_element_from,
+    scipy_sparse_matrix_from_w,
+    separate_components,
+    w_from_gdf,
+    array_from_graph_or_dict,
+    scipy_sparse_matrix_from_dict,
+)
 
 
 class AZP(BaseSpOptHeuristicSolver):
-    """ AZP involves class offering the implementation of the automatic zoning procedure algorithm.
-    
-        Parameters
-        ----------
-        gdf : geopandas.GeoDataFrame, required
-            Geodataframe containing original data
+    """AZP involves class offering the implementation of
+    the automatic zoning procedure algorithm.
 
-        w : libpysal.weights.W, required
-            Weights object created from given data
+    Parameters
+    ----------
+    gdf : geopandas.GeoDataFrame
+        Geodataframe containing original data.
 
-        attrs_name : list, required
-            Strings for attribute names (cols of ``geopandas.GeoDataFrame``).
-        
-        n_clusters : int, optional, default: 5
-            The number of clusters to form.
+    w : libpysal.weights.W
+        Weights object created from given data.
 
-        allow_move_strategy : None or :class:`AllowMoveStrategy`, default: None
-            For a different behavior for allowing moves an AllowMoveStrategy
-            instance can be passed as argument.
+    attrs_name : list
+        Strings for attribute names (cols of ``geopandas.GeoDataFrame``).
 
-        random_state : None, int, str, bytes, or bytearray, default: None
-            Random seed.
+    n_clusters : int
+        The number of clusters to form. Default is ``5``.
 
-        initial_labels : :class:`numpy.ndarray` or None, default: None
-            One-dimensional array of labels at the beginning of the algorithm.
-            If None, then a random initial clustering will be generated.
-            
-        objective_func : : class:`spopt.region.objective_function.ObjectiveFunction`, default: ObjectiveFunctionPairwise()
-            The objective function to use.
+    allow_move_strategy : None or AllowMoveStrategy
+        For a different behavior for allowing moves an AllowMoveStrategy
+        instance can be passed as argument. Default is ``None``.
+
+    random_state : None, int, str, bytes, or bytearray
+        Random seed. Default is ``None``.
+
+    initial_labels : numpy.ndarray or None
+        One-dimensional array of labels at the beginning of the algorithm.
+        If ``None``, then a random initial clustering will be generated.
+        Default is ``None``.
+
+    objective_func : spopt.region.objective_function.ObjectiveFunction
+        The objective function to use. Default is
+        ``ObjectiveFunctionPairwise()``.
+
[…]
