### 2021-06-19 a491568
blacken and clean azp_util
