+++ spopt/tests/test_maxp.py
@@ -1,8 +1,6 @@
-import numpy
+import geopandas
 import libpysal
-import geopandas as gpd
-import os
-import sys
+import numpy
 import unittest
 
 from spopt.region import MaxPHeuristic
@@ -10,7 +8,7 @@ from spopt.region import MaxPHeuristic
 
 # Mexican states
 pth = libpysal.examples.get_path("mexicojoin.shp")
-MEXICO = gpd.read_file(pth)
+MEXICO = geopandas.read_file(pth)
 
 
 class TestMaxPHeuristic(unittest.TestCase):
@@ -22,11 +20,6 @@ class TestMaxPHeuristic(unittest.TestCase):
         # count=1, threshold=4, top_n=2
         self.basic_labels = [5, 5, 3, 6, 6, 6, 1, 1, 1, 1, 8, 6, 8, 2, 2, 8]
         self.basic_labels += [2, 8, 7, 7, 2, 7, 5, 3, 3, 5, 3, 4, 4, 4, 4, 7]
-
-        # labels for:
-        # count=4, threshold=10, top_n=5
-        self.complex_labels = [3, 3, 2, 10, 10, 2, 6, 9, 1, 6, 6, 10, 7, 4, 4, 7]
-        self.complex_labels += [4, 7, 1, 1, 9, 1, 3, 8, 8, 2, 8, 2, 5, 5, 5, 9]
 
         # labels for one variable column:
         # count=1, threshold=5, top_n=5
@@ -45,7 +38,7 @@ class TestMaxPHeuristic(unittest.TestCase):
         model = MaxPHeuristic(*args)
         model.solve()
 
-        numpy.array_equal(model.labels_, self.basic_labels)
+        numpy.testing.assert_array_equal(model.labels_, self.basic_labels)
 
     def test_maxp_heuristic_basic_verbose(self):
         self.mexico["count"] = 1
@@ -59,9 +52,13 @@ class TestMaxPHeuristic(unittest.TestCase):
         model = MaxPHeuristic(*args, verbose=True)
         model.solve()
 
-        numpy.array_equal(model.labels_, self.basic_labels)
+        numpy.testing.assert_array_equal(model.labels_, self.basic_labels)
 
     def test_maxp_heuristic_complex(self):
+        """Can't set seed for MaxP effectively, therefore, label test fails.
+        Test for number of regions in order to hit lines of simulated
+        annealling code, etc.
+        """
         self.mexico["count"] = 4
         w = libpysal.weights.Queen.from_dataframe(self.mexico)
         attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
@@ -73,7 +70,7 @@ class TestMaxPHeuristic(unittest.TestCase):
         model = MaxPHeuristic(*args)
         model.solve()
 
-        numpy.array_equal(model.labels_, self.complex_labels)
+        self.assertEqual(model.p, threshold)
 
     def test_maxp_one_var(self):
         self.mexico["count"] = 1
@@ -87,4 +84,4 @@ class TestMaxPHeuristic(unittest.TestCase):
         model = MaxPHeuristic(*args)
         model.solve()
 
-        numpy.array_equal(model.labels_, self.var1_labels)
+        numpy.testing.assert_array_equal(model.labels_, self.var1_labels)
