+++ spopt/tests/test_maxp.py
@@ -21,6 +21,11 @@ class TestMaxPHeuristic(unittest.TestCase):
         self.basic_labels = [5, 5, 3, 6, 6, 6, 1, 1, 1, 1, 8, 6, 8, 2, 2, 8]
         self.basic_labels += [2, 8, 7, 7, 2, 7, 5, 3, 3, 5, 3, 4, 4, 4, 4, 7]
 
+        # labels for:
+        # count=4, threshold=10, top_n=5
+        self.complex_labels = [3, 3, 2, 10, 10, 2, 5, 6, 1, 7, 7, 10, 7, 4, 4]
+        self.complex_labels += [6, 4, 6, 1, 1, 9, 9, 3, 8, 8, 2, 2, 2, 5, 8, 5, 9]
+
         # labels for one variable column:
         # count=1, threshold=5, top_n=5
         self.var1_labels = [3, 3, 6, 2, 2, 2, 2, 4, 2, 4, 5, 2, 5, 1, 1, 5, 1]
@@ -55,10 +60,6 @@ class TestMaxPHeuristic(unittest.TestCase):
         numpy.testing.assert_array_equal(model.labels_, self.basic_labels)
 
     def test_maxp_heuristic_complex(self):
-        """Can't set seed for MaxP effectively, therefore, label test fails.
-        Test for number of regions in order to hit lines of simulated
-        annealling code, etc.
-        """
         self.mexico["count"] = 4
         w = libpysal.weights.Queen.from_dataframe(self.mexico)
         attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
@@ -70,7 +71,7 @@ class TestMaxPHeuristic(unittest.TestCase):
         model = MaxPHeuristic(*args)
         model.solve()
 
-        self.assertEqual(model.p, threshold)
+        numpy.testing.assert_array_equal(model.labels_, self.complex_labels)
 
     def test_maxp_one_var(self):
         self.mexico["count"] = 1
