### 2021-06-20 67ca41a
remove test azp sim. annealing
