+++ spopt/region/random_region.py
@@ -5,7 +5,7 @@ Randomly form regions given various types of constraints on cardinality and
 composition.
 """
 
-__author__ = "David Folch dfolch@fsu.edu, Serge Rey srey@asu.edu"
+__author__ = "David Folch David.Folch@nau.edu, Serge Rey sergio.rey@ucr.edu"
 
 import numpy as np
 import copy
@@ -143,14 +143,32 @@ class RandomRegions:
     [62, 61, 71, 63]
 
     """
+
     def __init__(
-        self, area_ids, num_regions=None, cardinality=None, contiguity=None,
-        maxiter=100, compact=False, max_swaps=1000000, permutations=99):
+        self,
+        area_ids,
+        num_regions=None,
+        cardinality=None,
+        contiguity=None,
+        maxiter=100,
+        compact=False,
+        max_swaps=1000000,
+        permutations=99,
+    ):
 
         solutions = []
         for i in range(permutations):
-            solutions.append(RandomRegion(area_ids, num_regions, cardinality,
-                                           contiguity, maxiter, compact, max_swaps))
+            solutions.append(
+                RandomRegion(
+                    area_ids,
+                    num_regions,
+                    cardinality,
+                    contiguity,
+                    maxiter,
+                    compact,
+                    max_swaps,
+                )
+            )
         self.solutions = solutions
         self.solutions_feas = []
         for i in solutions:
@@ -286,9 +304,17 @@ class RandomRegion:
     [37, 36, 38, 39]
 
     """
+
     def __init__(
-        self, area_ids, num_regions=None, cardinality=None, contiguity=None,
-                    maxiter=1000, compact=False, max_swaps=1000000):
+        self,
+        area_ids,
+        num_regions=None,
+        cardinality=None,
+        contiguity=None,
+        maxiter=1000,
+        compact=False,
+        max_swaps=1000000,
+    ):
 
         self.n = len(area_ids)
         ids = copy.copy(area_ids)
@@ -301,21 +327,22 @@ class RandomRegion:
         if cardinality:
             if self.n != sum(cardinality):
                 self.feasible = False
-                raise Exception('number of areas does not match cardinality')
+                raise Exception("number of areas does not match cardinality")
         if contiguity:
             if area_ids != contiguity.id_order:
                 self.feasible = False
-                raise Exception('order of area_ids must match order in contiguity')
+                raise Exception("order of area_ids must match order in contiguity")
         if num_regions and cardinality:
             if num_regions != len(cardinality):
                 self.feasible = False
-                raise Exception('number of regions does not match cardinality')
+                raise Exception("number of regions does not match cardinality")
 
         # dispatches the appropriate algorithm
         if num_regions and cardinality and contiguity:
             # conditioning on cardinality and contiguity (number of regions implied)
-            self.build_contig_regions(num_regions, cardinality, contiguity,
-                                      maxiter, compact, max_swaps)
+            self.build_contig_regions(
+                num_regions, cardinality, contiguity, maxiter, compact, max_swaps
+            )
         elif num_regions and cardinality:
             # conditioning on cardinality (number of regions implied)
             region_breaks = self.cards2breaks(cardinality)
@@ -323,13 +350,15 @@ class RandomRegion:
         elif num_regions and contiguity:
             # conditioning on number of regions and contiguity
             cards = self.get_cards(num_regions)
-            self.build_contig_regions(num_regions, cards, contiguity,
-                                      maxiter, compact, max_swaps)
+            self.build_contig_regions(
+                num_regions, cards, contiguity, maxiter, compact, max_swaps
+            )
         elif cardinality and contiguity:
             # conditioning on cardinality and contiguity
             num_regions = len(cardinality)
[…]
