### 2021-06-21 cf81c45
compact RandomRegion test empir
