### 2021-06-21 bf1286e
improve err msgs and test RandomRegion
