### 2021-06-21 70ecdf2
non-default Skater tests
