### 2021-06-21 1ece714
blacken regionkmeans, clean up tests
