+++ spopt/tests/test_region_k_means.py
@@ -1,17 +1,24 @@
-import numpy
 import libpysal
-import os
-import sys
+import numpy
+import unittest
 
-# sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), ".")))
 from spopt.region import RegionKMeansHeuristic
 
 
-def test_RegionKMeansHeuristic():
-    numpy.random.seed(12345)
-    w = libpysal.weights.lat2W(3, 3)
-    data = numpy.random.normal(size=(w.n, 3))
-    RKM = RegionKMeansHeuristic
-    model = RKM(data, 3, w)
-    model.solve()
-    numpy.array_equal(model.labels_, numpy.array([1, 2, 2, 1, 2, 0, 1, 1, 0]))
+RANDOM_STATE = 12345
+
+
+class TestSkater(unittest.TestCase):
+    def setUp(self):
+
+        self.dim = 3
+        self.w = libpysal.weights.lat2W(self.dim, self.dim)
+        self.data = numpy.random.normal(size=(self.w.n, self.dim))
+        self.known_labels = [0, 1, 1, 1, 0, 0, 2, 2, 2]
+
+    def test_region_k_means_heuristic(self):
+        numpy.random.seed(RANDOM_STATE)
+        model = RegionKMeansHeuristic(self.data, self.dim, self.w)
+        model.solve()
+
+        numpy.testing.assert_equal(model.labels_, self.known_labels)
