+++ spopt/region/ward.py
@@ -8,85 +8,87 @@ from sklearn.cluster import (
     SpectralClustering,
 )
 
+
 class WardSpatial(BaseSpOptHeuristicSolver):
-    """ Agglomerative clustering using Ward linkage with a spatial connectivity constraint.
-        
-        Parameters
-        ----------
-        
-        gdf : geopandas.GeoDataFrame, required
-            Geodataframe containing original data
+    """Agglomerative clustering using Ward linkage with a spatial connectivity constraint.
 
-        w : libpysal.weights.W, required
-            Weights object created from given data
+    Parameters
+    ----------
 
-        attrs_name : list, required
-            Strings for attribute names (cols of ``geopandas.GeoDataFrame``).
+    gdf : geopandas.GeoDataFrame, required
+        Geodataframe containing original data
 
-        n_clusters : int, optional, default: 5
-            The number of clusters to form.
+    w : libpysal.weights.W, required
+        Weights object created from given data
 
-        clustering_kwds: dictionary, optional, default: dict()
-            Other parameters about clustering could be used in sklearn.cluster.AgglometariveClustering.
-        
-        Returns
-        -------
+    attrs_name : list, required
+        Strings for attribute names (cols of ``geopandas.GeoDataFrame``).
 
-        labels_ : numpy.array
-            Cluster labels for observations.
+    n_clusters : int, optional, default: 5
+        The number of clusters to form.
+
+    clustering_kwds: dictionary, optional, default: dict()
+        Other parameters about clustering could be used in sklearn.cluster.AgglometariveClustering.
+
+    Returns
+    -------
+
+    labels_ : numpy.array
+        Cluster labels for observations.
 
 
-        Examples
-        --------
+    Examples
+    --------
 
-        >>> import numpy as np
-        >>> import libpysal
-        >>> import geopandas as gpd
-        >>> from spopt.region import WardSpatial
+    >>> import numpy as np
+    >>> import libpysal
+    >>> import geopandas as gpd
+    >>> from spopt.region import WardSpatial
 
-        Read the data.
+    Read the data.
 
-        >>> pth = libpysal.examples.get_path('airbnb_Chicago 2015.shp')
-        >>> chicago = gpd.read_file(pth)
+    >>> pth = libpysal.examples.get_path('airbnb_Chicago 2015.shp')
+    >>> chicago = gpd.read_file(pth)
 
-        Initialize the parameters.
+    Initialize the parameters.
 
-        >>> w = libpysal.weights.Queen.from_dataframe(chicago)
-        >>> attrs_name = ['num_spots']
-        >>> n_clusters = 8
+    >>> w = libpysal.weights.Queen.from_dataframe(chicago)
+    >>> attrs_name = ['num_spots']
+    >>> n_clusters = 8
 
-        Run the skater algorithm.
+    Run the skater algorithm.
 
-        >>> model = WardSpatial(chicago, w, attrs_name, n_clusters)
-        >>> model.solve()
+    >>> model = WardSpatial(chicago, w, attrs_name, n_clusters)
+    >>> model.solve()
 
-        Get the region IDs for unit areas.
+    Get the region IDs for unit areas.
 
-        >>> model.labels_
-        
-        Show the clustering results.
+    >>> model.labels_
 
-        >>> chicago['ward_new'] = model.labels_
-        >>> chicago.plot(column='ward_new', categorical=True, figsize=(12,8), edgecolor='w')
+    Show the clustering results.
 
-        """
+    >>> chicago['ward_new'] = model.labels_
+    >>> chicago.plot(column='ward_new', categorical=True, figsize=(12,8), edgecolor='w')
 
+    """
 
[…]
