### 2021-06-22 #3330 @vissarion
Hi, @nickbearman, @ryankarlos, @ystouthart I think this submission matches your interests.

Would any of you be willing to review this submission for JOSS?
