### 2021-06-22 #3330 @nickbearman
Yes, happy to review. Just to check, I am not actually going to be able to run the package to test it out (not currently having the time/space to setup a Python environment), is this a major issue? If so, happy to pass the review to someone else. Very happy to complete the rest of the review elements.
