### 2021-06-22 #3330 @vissarion
Hi @nickbearman thank you very much for you interest and quick response!

I am afraid it is important to test the functionality of the package as stated in https://joss.readthedocs.io/en/latest/review_criteria.html

> Reviewers are expected to install the software they are reviewing and to verify the core functionality of the software.
