### 2021-06-22 18e1ed2
solver tests in util
