### 2021-06-22 13566df
update pulp api check for solvers
