### 2021-06-22 dbf69b6
chore: add test for coverage model problems with synthetic data
