+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos: 
 -   repo: https://github.com/ambv/black
-    rev: stable
+    rev: 21.6b0
     hooks: 
     - id: black
       language_version: python3.8
