### 2021-06-22 9da7546
chore: update .ci folder adding spaghetti and shape as dep
