+++ .ci/38.yaml
@@ -12,7 +12,9 @@ dependencies:
   - pulp
   - scikit-learn>=0.22
   - scipy>=1.0
+  - shapely
   - spenc
+  - spaghetti
   # testing, docs, etc.
   - codecov
   - coverage
