### 2021-06-22 8cdb886
fix: add installation of glpk in github actions
