### 2021-06-22 a4d0bf5
fix: change GLPK solver to COIN as default
