### 2021-06-22 5031488
fix: change GLPK solver to PULP_COIN as default
