+++ .github/workflows/unittests.yml
@@ -17,11 +17,9 @@
        matrix:
          os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
          environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
-         exclude:
-           - environment-file: .ci/39.yaml
-             os: windows-latest
-           #- environment-file: .ci/38.yaml
-           #  os: ubuntu-latest
+         #exclude:
+         #  - environment-file: .ci/39.yaml
+         #    os: windows-latest
          include:
            - environment-file: .ci/39-DEV.yaml
              os: ubuntu-latest
