+++ .github/workflows/unittests.yml
@@ -17,9 +17,6 @@
        matrix:
          os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
          environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
-         #exclude:
-         #  - environment-file: .ci/39.yaml
-         #    os: windows-latest
          include:
            - environment-file: .ci/39-DEV.yaml
              os: ubuntu-latest
