### 2021-06-24 #3330 @vissarion
Thanks @samuel-rosa for your recommendation.

Hi, @AlexandreWadoux, would you be willing to review this submission for JOSS?
