### 2021-06-24 #3330 @vissarion
@AlexandreWadoux I understand your concern. Thanks for the quick reply!

As a side note, in JOSS a submission is accepted only if all reviewers recommend acceptance according to those criteria https://joss.readthedocs.io/en/latest/review_criteria.html
