+++ spopt/tests/test_skater.py
@@ -7,56 +7,77 @@ import unittest
 from spopt.region import Skater
 
 
-# Empirical tests -- Mexican states
+# Empirical tests
 RANDOM_STATE = 12345
+# -- Mexican states
 pth = libpysal.examples.get_path("mexicojoin.shp")
 MEXICO = geopandas.read_file(pth)
+# -- Columbus
+pth = libpysal.examples.get_path("columbus.shp")
+COLUMBUS = geopandas.read_file(pth)
 
 
 class TestSkater(unittest.TestCase):
     def setUp(self):
-        self.mexico = MEXICO.copy()
-        self.w = libpysal.weights.Queen.from_dataframe(self.mexico)
-        self.default_attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
-        self.default_labels = [0, 0, 1, 2, 2, 1, 1, 1, 1, 1, 3, 2, 1, 1, 1, 1]
-        self.default_labels += [4, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1]
 
-        self.non_default_labels = [0, 0, 1, 2, 2, 3, 4, 5, 6, 4, 4, 2, 5, 7, 7, 5, 7]
-        self.non_default_labels += [5, 6, 6, 6, 6, 0, 8, 8, 1, 1, 3, 3, 8, 3, 3]
+        # Mexico
+        self.mexico = MEXICO.copy()
+        self.w_mexico = libpysal.weights.Queen.from_dataframe(self.mexico)
+        self.default_attrs_mexico = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
+        self.default_mexico = [0, 0, 1, 2, 2, 1, 1, 1, 1, 1, 3, 2, 1, 1, 1, 1]
+        self.default_mexico += [4, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 0, 1, 1]
+        self.non_default_mexico = [0, 0, 1, 2, 2, 3, 4, 5, 6, 4, 4, 2, 5, 7, 7, 5, 7]
+        self.non_default_mexico += [5, 6, 6, 6, 6, 0, 8, 8, 1, 1, 3, 3, 8, 3, 3]
+
+        # Columbus (synth island)
+        self.columbus = COLUMBUS.copy()
+        remove = [13, 14, 17, 18, 20, 23, 24, 29]
+        self.columbus = self.columbus[~self.columbus.index.isin(remove)]
+        self.w_columbus = libpysal.weights.Queen.from_dataframe(self.columbus)
+        self.attrs_columbus = ["HOVAL", "INC", "CRIME", "OPEN", "PLUMB", "DISCBD"]
+        self.columbus_labels = [0, 0, 0, 0, 1, 1, 0, 0, 2, 3, 1, 1, 0, 1, 3]
+        self.columbus_labels += [3, 2, 3, 2, 2, 2, 2, 4, 3, 2, 4, 2, 4]
+        self.columbus_labels += [5, 2, 4, 3, 3, 4, 5, 5, 5, 4, 3, 5, 5]
 
     def test_skater_defaults(self):
         self.mexico["count"] = 1
         numpy.random.seed(RANDOM_STATE)
-        model = Skater(gdf=self.mexico, w=self.w, attrs_name=self.default_attrs_name)
+        model = Skater(self.mexico, self.w_mexico, self.default_attrs_mexico)
         model.solve()
 
-        numpy.testing.assert_equal(model.labels_, self.default_labels)
+        numpy.testing.assert_equal(model.labels_, self.default_mexico)
 
     def test_skater_defaults_verbose(self):
         self.mexico["count"] = 1
         sfkws = {"verbose": 1}
[…]
