### 2021-06-24 1a5f21f
add spenc non-defaults test
