+++ spopt/tests/test_spenc.py
@@ -1,22 +1,53 @@
-import numpy as np
+import geopandas
 import libpysal
-import geopandas as gpd
-from spopt.region import Spenc
+import numpy
 import unittest
+
+from spopt.region import Spenc
+
+
+# Empirical tests
+RANDOM_STATE = 123456
+# -- Mexican states
+pth = libpysal.examples.get_path("mexicojoin.shp")
+MEXICO = geopandas.read_file(pth)
 
 
 class TestSpenc(unittest.TestCase):
     def setUp(self):
-        self.pth = libpysal.examples.get_path("mexicojoin.shp")
-        self.mexico = gpd.read_file(self.pth)
-        self.mexico["count"] = 1
-        self.attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
-        self.w = libpysal.weights.Queen.from_dataframe(self.mexico)
-        self.model = Spenc(
-            gdf=self.mexico, w=self.w, attrs_name=self.attrs_name, random_state=123456
-        )
-        self.model.solve()
 
-    def test_labels_(self):
-        observed_labels = self.model.labels_
-        np.testing.assert_equal(len(np.unique(observed_labels)), 5)
+        # Mexico
+        self.mexico = MEXICO.copy()
+        self.w_mexico = libpysal.weights.Queen.from_dataframe(self.mexico)
+        self.default_attrs_mexico = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
+        self.default_mexico = [3, 0, 3, 4, 4, 3, 4, 1, 4, 2, 1, 1, 1, 3, 3, 3]
+        self.default_mexico += [0, 2, 0, 2, 2, 0, 0, 0, 0, 0, 3, 4, 1, 2, 0, 3]
+
+        self.non_default_mexico = [0, 0, 0, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 1, 2]
+        self.non_default_mexico += [1, 2, 2, 2, 1, 2, 0, 0, 0, 0, 0, 2, 2, 2, 2, 2]
+
+    def test_spenc_defaults(self):
+        self.mexico["count"] = 1
+        model = Spenc(
+            self.mexico,
+            self.w_mexico,
+            self.default_attrs_mexico,
+            random_state=RANDOM_STATE,
+        )
+        model.solve()
+
+        numpy.testing.assert_equal(model.labels_, self.default_mexico)
+
+    def test_spenc_non_defaults(self):
+        self.mexico["count"] = 1
+        model = Spenc(
+            self.mexico,
+            self.w_mexico,
+            self.default_attrs_mexico,
+            random_state=RANDOM_STATE,
+            gamma=0,
+            n_clusters=3,
+        )
+        model.solve()
+
+        numpy.testing.assert_equal(model.labels_, self.non_default_mexico)
