### 2021-06-24 7dbe6a5
add spenc non-defaults test [2]
