### 2021-06-24 7ae3ff2
add spenc non-defaults test [3]
