### 2021-06-24 fbd67f2
add spenc non-defaults test [3]
