### 2021-06-25 81f3dca
add complex test for region-kmeans
