+++ .github/workflows/release_and_publish.yml
@@ -1,4 +1,3 @@
-
 name: Release Package
 
 on:
@@ -16,49 +15,42 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: '3.x'
+    
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install setuptools wheel twine jupyter urllib3 pandas pyyaml
         python setup.py sdist bdist_wheel
-    - name: Publish distribution 📦 to PyPI
-      uses: pypa/gh-action-pypi-publish@master
-      with:
-        user: __token__
-        password: ${{ secrets.pypi_password }}
-    - name: Run Changelog
+    
+    - name: run Changelog
       run: |
         jupyter nbconvert --to notebook --execute --inplace --ExecutePreprocessor.timeout=-1 --ExecutePreprocessor.kernel_name=python3 tools/gitcount.ipynb
-    - name: Cat Changelog
+    
+    - name: cat Changelog
       uses: pCYSl5EDgo/cat@master
       id: changetxt
       with:
         path: ./tools/changelog.md
       env:
         TEXT: ${{ steps.changetxt.outputs.text }}
-    - name: Create Release
-      id: create_release
-      uses: actions/create-release@v1
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # This token is provided by Actions, you do not need to create your own token
-      with:
-        tag_name: ${{ github.ref }}
-        release_name: Release ${{ github.ref }}
-        body: ${{ steps.changetxt.outputs.text }}
-        draft: false
-        prerelease: false
-    - name: Get Asset name
-      run: |
-        export PKG=$(ls dist/)
-        set -- $PKG
-        echo "name=$1" >> $GITHUB_ENV
-    - name: Upload Release Asset
-      id: upload-release-asset 
-      uses: actions/upload-release-asset@v1
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      with:
-        upload_url: ${{ steps.create_release.outputs.upload_url }} # This pulls from the CREATE RELEASE step above, referencing it's ID to get its outputs object, which include a `upload_url`. See this blog post for more info: https://jasonet.co/posts/new-features-of-github-actions/#passing-data-to-future-steps 
-        asset_path: dist/${{ env.name }}
-        asset_name: ${{ env.name }}
-        asset_content_type: application/zip
+    
+   - name: Get the tag name
+     run: echo "TAG=${GITHUB_REF/refs\/tags\//}" >> $GITHUB_ENV
+  
+   - name: Release
+     uses: softprops/action-gh-release@v1
+     with:
+       body: ${{ steps.changetxt.outputs.text }}
+       body_path: ${{ steps.changetxt.outputs.path }}
+       name: Release ${{ env.TAG }}
+     env:
+       GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+  
+   - name: Publish distribution 📦 to PyPI
+     uses: pypa/gh-action-pypi-publish@master
+     with:
+       user: __token__
+       password: ${{ secrets.pypi_password }}
+      
+      
+      
