+++ .github/workflows/build_docs.yml
@@ -19,7 +19,7 @@
      strategy:
        matrix:
          os: ['ubuntu-latest']
-         environment-file: [ci/39.yaml]
+         environment-file: [.ci/39.yaml]
          experimental: [false]
      defaults:
        run:
