+++ docs/conf.py
@@ -280,8 +280,8 @@ numpydoc_xref_ignore = {
 
 # This is processed by Jinja2 and inserted before each notebook
 nbsphinx_prolog = r"""
-{% set docname = env.doc2path(env.docname, base='').replace('nblink','ipynb') %}
-{% set fullpath = env.doc2path(env.docname, base='tree/main/').replace('nblink','ipynb') %}
+{% set docname = env.doc2path(env.docname, base='') %}
+{% set fullpath = env.doc2path(env.docname, base='tree/main/') %}
 
 .. only:: html
 
