+++ spopt/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.1.1"
+__version__ = "0.1.2"
 
 from . import region
