### 2021-06-28 7d4fd62
chore: add euclidean and manhattan distance support
