+++ spopt/locate/coverage.py
@@ -5,12 +5,46 @@ import pulp
 from geopandas import GeoDataFrame
 from abc import ABC, abstractmethod
 from spopt.locate.base import LocateSolver, FacilityModelBuilder
+from scipy.spatial import distance_matrix
 
 
-class LSCP(LocateSolver):
+class Coverage:
     def __init__(self, name: str, problem: pulp.LpProblem):
         self.name = name
         self.problem = problem
+        self.aij = np.array([[]])
+
+    def uncovered_clients_dict(self):
+        self.n_cli_uncov = self.aij.shape[0] - len(self.cli2iloc.keys())
+
+    def client_facility_dict(self):
+        self.cli2fac = {}
+        for cv in list(self.cli2iloc.keys()):
+            self.cli2fac[cv] = []
+            for k, v in self.fac2cli.items():
+                if cv in v:
+                    self.cli2fac[cv].append(k)
+
+    def cov_dict(self):
+        self.cli2ncov = {}
+        for c, fs in self.cli2fac.items():
+            self.cli2ncov[c] = len(fs)
+            most_coverage = max(self.cli2ncov.values())
+            self.ncov2ncli = {}
+            for cov_count in range(most_coverage + 1):
+                if cov_count == 0:
+                    self.ncov2ncli[cov_count] = self.n_cli_uncov
+                    continue
+                if not cov_count in list(self.cli2ncov.keys()):
+                    self.ncov2ncli[cov_count] = 0
+                for c, ncov in self.cli2ncov.items():
+                    if ncov >= cov_count:
+                        self.ncov2ncli[cov_count] += 1
+
+
+class LSCP(LocateSolver, Coverage):
+    def __init__(self, name: str, problem: pulp.LpProblem):
+        super().__init__(name, problem)
 
     def __add_obj(self):
         fac_vars = getattr(self, "fac_vars")
@@ -25,14 +59,14 @@ class LSCP(LocateSolver):
         model = pulp.LpProblem(name, pulp.LpMinimize)
         lscp = LSCP(name, model)
 
-        FacilityModelBuilder.add_facility_integer_variable(lscp, r_fac)
+        FacilityModelBuilder.add_facility_integer_variable(lscp, r_fac, "x[{i}]")
 
-        aij = np.zeros(cost_matrix.shape)
-        aij[cost_matrix <= max_coverage] = 1
+        lscp.aij = np.zeros(cost_matrix.shape)
+        lscp.aij[cost_matrix <= max_coverage] = 1
 
         lscp.__add_obj()
         FacilityModelBuilder.add_set_covering_constraint(
-            lscp, lscp.problem, aij, r_fac, r_cli
+            lscp, lscp.problem, lscp.aij, r_fac, r_cli
         )
 
         return lscp
@@ -47,26 +81,46 @@ class LSCP(LocateSolver):
         max_coverage: float,
         distance_metric: str = "euclidean",
     ):
+        dem = gdf_demand[demand_col]
+        fac = gdf_fac[facility_col]
+
+        dem_data = np.array([dem.x.to_numpy(), dem.y.to_numpy()]).T
+        fac_data = np.array([fac.x.to_numpy(), fac.y.to_numpy()]).T
+
         distances = np.array([])
+
         if gdf_demand.crs != gdf_fac.crs:
             raise ValueError(
                 f"geodataframes crs are different: gdf_demand-{gdf_demand.crs}, gdf_fac-{gdf_fac.crs}"
             )
 
         if distance_metric == "manhattan":
-            # TODO
-            pass
+            distances = distance_matrix(dem_data, fac_data, p=1)
         elif distance_metric == "euclidean":
-            distances = gdf_demand[demand_col].apply(
-                lambda geometry: gdf_fac[facility_col].distance(geometry)
[…]
