+++ _sources/tutorials.rst.txt
@@ -1,0 +1,13 @@
+Tutorials
+=========
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Regionalization:
+
+    notebooks/maxp.ipynb
+    notebooks/azp.ipynb
+    notebooks/skater.ipynb
+    notebooks/ward.ipynb
+    notebooks/reg-k-means.ipynb
+    notebooks/randomregion.ipynb
