+++ _sources/api.rst.txt
@@ -1,0 +1,27 @@
+.. _api_ref:
+
+.. currentmodule:: spopt
+
+API reference
+=============
+
+.. _data_api:
+
+Region Methods
+-----------------------
+Model based approaches for aggregating a large set of geographic units (with small footprints) into a smaller number of regions (with large footprints).
+
+
+.. autosummary::
+   :toctree: generated/
+
+    region.AZP
+    region.MaxPHeuristic
+    region.RandomRegion
+    region.RandomRegions
+    region.RegionKMeansHeuristic
+    region.Skater
+    region.Spenc
+    region.WardSpatial
+
+
