+++ _sources/generated/spopt.region.WardSpatial.rst.txt
@@ -1,0 +1,23 @@
+﻿spopt.region.WardSpatial
+========================
+
+.. currentmodule:: spopt.region
+
+.. autoclass:: WardSpatial
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~WardSpatial.__init__
+      ~WardSpatial.solve
+   
+   
+
+   
+   
+   
