+++ spopt/locate/__init__.py
@@ -1,1 +1,3 @@
 from .coverage import LSCP, MCLP
+from .p_median import PMedian
+from .p_center import PCenter
