### 2021-07-05 b206357
chore: add docstrings to base.py [WIP]
