+++ spopt/locate/base.py
@@ -1,13 +1,29 @@
-from abc import ABC, abstractmethod, abstractproperty
-import pulp
-from geopandas import GeoDataFrame
+from abc import ABC, abstractmethod
+from typing import TypeVar
+
 import numpy as np
-from typing import Type, TypeVar
+import pulp
 
 
 class LocateSolver(ABC):
+    """
+    Base Class for locate package
+    """
+
     @abstractmethod
     def solve(self, solver: pulp.LpSolver):
+        """
+        Solve the optimization model
+
+        Parameters
+        ----------
+        solver: pulp.LpSolver
+                solver supported by pulp package
+
+        Returns
+        -------
+        None
+        """
         pass
 
 
@@ -16,11 +32,29 @@ T_FacModel = TypeVar("T_FacModel", bound=LocateSolver)
 
 class FacilityModelBuilder(object):
     """
-    Set Constraints and Variables
+    Set facility locations' variables and constraints
     """
 
     @staticmethod
-    def add_facility_integer_variable(obj: T_FacModel, range_facility, var_name):
+    def add_facility_integer_variable(
+        obj: T_FacModel, range_facility, var_name
+    ) -> None:
+        """
+
+        Parameters
+        ----------
+        obj: T_FacModel
+            bounded type of LocateSolver class
+        range_facility: range
+            range of facility points quantity
+        var_name: str
+            formatted string
+            facility variable name
+
+        Returns
+        -------
+        None
+        """
         fac_vars = [
             pulp.LpVariable(
                 var_name.format(i=i), lowBound=0, upBound=1, cat=pulp.LpInteger
@@ -31,7 +65,23 @@ class FacilityModelBuilder(object):
         setattr(obj, "fac_vars", fac_vars)
 
     @staticmethod
-    def add_client_integer_variable(obj: T_FacModel, range_client, var_name):
+    def add_client_integer_variable(obj: T_FacModel, range_client, var_name) -> None:
+        """
+
+        Parameters
+        ----------
+        obj: T_FacModel
+            bounded type of LocateSolver class
+        range_client: range
+            range of demand points quantity
+        var_name: str
+            formatted string
+            demand variable name
+
+        Returns
+        -------
+        None
+        """
         cli_vars = [
             pulp.LpVariable(
                 var_name.format(i=i), lowBound=0, upBound=1, cat=pulp.LpInteger
@@ -44,7 +94,25 @@ class FacilityModelBuilder(object):
     @staticmethod
     def add_client_assign_integer_variable(
         obj: T_FacModel, range_client, range_facility, var_name
-    ):
+    ) -> None:
+        """
+
+        Parameters
+        ----------
+        obj: T_FacModel
+            bounded type of LocateSolver class
+        range_client: range
+            range of demand points quantity
+        range_facility: range
+            range of facility points quantity
+        var_name: str
+            formatted string
+            client assigning variable name
+
+        Returns
+        -------
+        None
+        """
         cli_assgn_vars = [
             [
                 pulp.LpVariable(
@@ -58,15 +126,52 @@ class FacilityModelBuilder(object):
         setattr(obj, "cli_assgn_vars", cli_assgn_vars)
 
     @staticmethod
-    def add_weight_continuous_variable(obj: T_FacModel):
+    def add_weight_continuous_variable(obj: T_FacModel) -> None:
+        """
+
+        Parameters
+        ----------
+        obj: T_FacModel
+            bounded type of LocateSolver class
+
+        Returns
+        -------
+        None
+        """
         weight_var = pulp.LpVariable("W", lowBound=0, cat=pulp.LpContinuous)
 
         setattr(obj, "weight_var", weight_var)
 
     @staticmethod
[…]

[…]
