### 2021-07-06 d0a43f9
chore: add docstrings to coverage.py [WIP]
