+++ spopt/locate/coverage.py
@@ -3,7 +3,6 @@ import numpy as np
 import pulp
 from geopandas import GeoDataFrame
 
-import spopt.locate
 from spopt.locate.base import LocateSolver, FacilityModelBuilder
 from scipy.spatial.distance import cdist
 
@@ -43,19 +42,57 @@ class Coverage:
 
 
 class LSCP(LocateSolver, Coverage):
+    """
+    LSCP class implements Location Set Covering optimization model and solve it.
+
+    Parameters
+    ----------
+    name: str
+        Problem name
+    problem: pulp.LpProblem
+        Pulp instance of optimization model that contains constraints, variables and objective function.
+
+    """
+
     def __init__(self, name: str, problem: pulp.LpProblem):
         super().__init__(name, problem)
 
     def __add_obj(self) -> None:
+        """
+        Add objective function to model:
+        x1 + x2 + x3 + x4 + x5 + ... + xj
+
+        Returns
+        -------
+        None
+        """
         fac_vars = getattr(self, "fac_vars")
         self.problem += pulp.lpSum(fac_vars), "objective function"
 
     @classmethod
-    def from_cost_matrix(cls, cost_matrix: np.array, max_coverage: float):
+    def from_cost_matrix(
+        cls, cost_matrix: np.array, max_coverage: float, name: str = "LSCP"
+    ):
+        """
+        Create a LSCP object based on cost matrix.
+
+        Parameters
+        ----------
+        cost_matrix: np.array
+            two-dimensional distance array between facility points and demand point
+        max_coverage: float
+            maximum acceptable service distance by problem
+        name: str, default="LSCP"
+            name of the problem
+
+        Returns
+        -------
+        LSCP object
+        """
+
         r_fac = range(cost_matrix.shape[1])
         r_cli = range(cost_matrix.shape[0])
 
-        name = "LSCP"
         model = pulp.LpProblem(name, pulp.LpMinimize)
         lscp = LSCP(name, model)
 
@@ -80,7 +117,34 @@ class LSCP(LocateSolver, Coverage):
         facility_col: str,
         max_coverage: float,
         distance_metric: str = "euclidean",
+        name: str = "LSCP",
     ):
+        """
+        Create a LSCP object based on geodataframes. Calculates the cost matrix between demand and facility,
+        and then uses from_cost_matrix method.
+
+        Parameters
+        ----------
+        gdf_demand: geopandas.GeoDataFrame
+            demand geodataframe with point geometry
+        gdf_fac: geopandas.GeoDataframe
+            facility geodataframe with point geometry
+        demand_col: str
+            demand geometry column name
+        facility_col: str
+            facility candidate sites geometry column name
+        max_coverage: float
+            maximum acceptable service distance by problem
+        distance_metric: str, default="euclidean"
+            metrics supported by :method: `scipy.spatial.distance.cdist` used for the distance calculations
+        name: str, default="LSCP"
+            name of the problem
+
+        Returns
+        -------
+        LSCP object
+        """
+
         dem = gdf_demand[demand_col]
         fac = gdf_fac[facility_col]
 
@@ -96,7 +160,7 @@ class LSCP(LocateSolver, Coverage):
 
         distances = cdist(dem_data, fac_data, distance_metric)
 
-        return cls.from_cost_matrix(distances, max_coverage)
+        return cls.from_cost_matrix(distances, max_coverage, name)
 
     def record_decisions(self):
         fac_vars = getattr(self, "fac_vars")
@@ -118,6 +182,18 @@ class LSCP(LocateSolver, Coverage):
         self.cov_dict()
 
     def solve(self, solver: pulp.LpSolver):
+        """
+        Solve the LSCP model
+
+        Parameters
+        ----------
+        solver: pulp.LpSolver
[…]
