### 2021-07-07 00bb564
chore: add docstrings to p_center and p_median classes
