### 2021-07-12 ead7c94
chore: add lscp and mclp notebooks
