### 2021-07-13 80ebbab
chore: add tests to optimality to locate package
