+++ spopt/locate/p_center.py
@@ -172,12 +172,4 @@ class PCenter(LocateSolver):
         PCenter object
         """
         self.problem.solve(solver)
-
-        if self.problem.status == pulp.constants.LpStatusUnbounded:
-            raise Exception("unbounded solution")
-        elif self.problem.status == pulp.constants.LpStatusNotSolved:
-            raise Exception("not possible to solve")
-        elif self.problem.status == pulp.constants.LpSolutionInfeasible:
-            raise Exception("infeasible solution")
-        elif self.problem.status == pulp.constants.LpSolutionOptimal:
-            return self
+        return self
