+++ spopt/tests/test_locate.py
@@ -1,12 +1,13 @@
 import numpy
 import geopandas
+import pandas
 import pulp
 import spaghetti
 from shapely.geometry import Point
 
-import spopt.locate
 from spopt.locate import LSCP, MCLP, PCenter, PMedian
 import unittest
+import os
 
 
 def simulated_geo_points(in_data, needed=20, seed=0, to_file=None):
@@ -30,7 +31,7 @@ def simulated_geo_points(in_data, needed=20, seed=0, to_file=None):
     return sim_pts
 
 
-class TestLocate(unittest.TestCase):
+class TestGlobalLocate(unittest.TestCase):
     def setUp(self) -> None:
         lattice = spaghetti.regular_lattice((0, 0, 10, 10), 9, exterior=True)
         ntw = spaghetti.Network(in_data=lattice)
@@ -129,3 +130,117 @@ class TestLocate(unittest.TestCase):
         )
         result = p_center.solve(pulp.PULP_CBC_CMD())
         self.assertIsInstance(result, PCenter)
+
+
+class TestOptimalLocate(unittest.TestCase):
+    def setUp(self) -> None:
+        dirpath = os.path.join(os.path.dirname(__file__), "./data/")
+        network_distance = pandas.read_csv(
+            dirpath + "SF_network_distance_candidateStore_16_censusTract_205_new.csv"
+        )
+
+        ntw_dist_piv = network_distance.pivot_table(
+            values="distance", index="DestinationName", columns="name"
+        )
+        facilities_name = ntw_dist_piv.columns
+        demand_name = network_distance[["DestinationName", "demand"]].drop_duplicates()
+
+        self.cost_matrix = ntw_dist_piv.to_numpy()
+
+        demand_points = pandas.read_csv(
+            dirpath + "SF_demand_205_centroid_uniform_weight.csv"
+        )
+        facility_points = pandas.read_csv(dirpath + "SF_store_site_16_longlat.csv")
+
+        self.facility_points_gdf = geopandas.GeoDataFrame(
+            facility_points,
+            geometry=geopandas.points_from_xy(
+                facility_points.long, facility_points.lat
+            ),
+        )
+        self.demand_points_gdf = geopandas.GeoDataFrame(
+            demand_points,
+            geometry=geopandas.points_from_xy(demand_points.long, demand_points.lat),
+        )
+
+        self.service_dist = 5000.0
+        self.p_facility = 4
+        self.ai = demand_name.sort_values(by=["DestinationName"])["demand"].to_numpy()
+
+    def test_lscp_from_cost_matrix(self):
+        lscp = LSCP.from_cost_matrix(self.cost_matrix, self.service_dist)
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+
+        self.assertEqual(lscp.problem.status, pulp.LpStatusOptimal)
+
+    def test_lscp_from_geodataframe(self):
+        lscp = LSCP.from_geodataframe(
+            self.demand_points_gdf,
+            self.facility_points_gdf,
+            "geometry",
+            "geometry",
+            self.service_dist,
+        )
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+        self.assertEqual(lscp.problem.status, pulp.LpStatusOptimal)
+
+    def test_mclp_from_cost_matrix(self):
+        mclp = MCLP.from_cost_matrix(
+            self.cost_matrix,
+            self.ai,
+            max_coverage=self.service_dist,
+            p_facilities=self.p_facility,
+        )
+        mclp = mclp.solve(pulp.PULP_CBC_CMD())
+        self.assertEqual(mclp.problem.status, pulp.LpStatusOptimal)
+
[…]
