### 2021-07-14 21c86d1
chore: add support to different geom type in locate package
