### 2021-07-15 674a5f6
chore: add infesibility tests to locate package
