### 2021-07-17 ad545e2
reformat citation in docs
