+++ docs/index.rst
@@ -17,7 +17,7 @@ Regionalization, facility location, and transportation-oriented modeling
       <div class="row equal-height">
         <div class="col-sm-.5 col-xs-hidden">
         </div>
-        <div class="col-md-3.5 col-xs-13">
+        <div class="col-md-4 col-xs-14">
             <a 
             href="https://pysal.org/spopt/notebooks/maxp.html" class="thumbnail">
                 <img src="_static/images/notebooks_maxp_12_1.png" class="img-responsive center-block">
@@ -26,7 +26,7 @@ Regionalization, facility location, and transportation-oriented modeling
                 </div>
             </a>
         </div>
-        <div class="col-sm-3.5 col-xs-13">
+        <div class="col-sm-4 col-xs-14">
             <a href="https://pysal.org/spopt/notebooks/skater.html" class="thumbnail">
                 <img src="_static/images/notebooks_skater_12_1.png" class="img-responsive center-block">
                 <div class="caption text-center">
@@ -34,7 +34,7 @@ Regionalization, facility location, and transportation-oriented modeling
                 </div>
             </a>
         </div>
-        <div class="col-sm-3.5 col-xs-13">
+        <div class="col-sm-4 col-xs-14">
             <a href="https://pysal.org/spopt/notebooks/reg-k-means.html" class="thumbnail">
                 <img src="_static/images/notebooks_reg-k-means_15_1.png"
                 class="img-responsive center-block">
