+++ .ci/39.yaml
@@ -21,6 +21,6 @@ dependencies:
   # docs
   - nbsphinx
   - numpydoc
-  - sphinx>=1.4.3
-  - sphinxcontrib-bibtex<2.0.0
+  - sphinx
+  - sphinxcontrib-bibtex
   - sphinx_bootstrap_theme
