+++ README.md
@@ -26,7 +26,7 @@ mexico = geopandas.read_file(libpysal.examples.get_path("mexicojoin.shp"))
 mexico["count"] = 1
 attrs = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
 w = libpysal.weights.Queen.from_dataframe(mexico)
-mexico["count"], threshold_name, threshold, top_n  = 1, "count", 4, 2
+mexico["count"], threshold_name, threshold, top_n = 1, "count", 4, 2
 numpy.random.seed(123456)
 model = spopt.MaxPHeuristic(mexico, w, attrs, threshold_name, threshold, top_n)
 model.solve()
