+++ .github/workflows/unittests.yml
@@ -47,7 +47,7 @@
          if: matrix.os == 'windows-latest'
        
        - name: codecov
-         uses: codecov/codecov-action@v1
+         uses: codecov/codecov-action@v2.0.1
          with:
            token: ${{ secrets.CODECOV_TOKEN }}
            file: ./coverage.xml
