### 2021-07-19 50388d7
doc: adding more explanation to locate notebooks
