### 2021-07-23 8ba754d
chore: move client-facility lookup method from coverage to base class of locate package
