### 2021-07-29 2fbec62
chore: big changes in storing results of locate models
