+++ spopt/locate/__init__.py
@@ -1,3 +1,4 @@
+from .base import BaseOutputMixin, CoveragePercentageMixin, MeanDistanceMixin
 from .coverage import LSCP, MCLP
 from .p_median import PMedian
 from .p_center import PCenter
