### 2021-08-04 d5fc900
docs: add real world problem notebook
