### 2021-08-05 15c4d6a
fix: add spaghetti install 39-DEV.yaml
