+++ .ci/39-DEV.yaml
@@ -13,6 +13,7 @@ dependencies:
   - scikit-learn>=0.22
   - scipy>=1.0
   - spenc
+  - spaghetti
   # testing
   - codecov
   - coverage
