### 2021-08-05 77f6c51
enh: remove pulp install in unittests.yml
