+++ .ci/39.yaml
@@ -9,6 +9,7 @@ dependencies:
   - networkx
   - numpy>=1.3
   - pandas>=1
+  - pulp
   - scikit-learn>=0.22
   - scipy>=1.0
   - shapely
