+++ .ci/39-DEV.yaml
@@ -9,6 +9,7 @@ dependencies:
   - numpy>=1.3
   - pandas>=1
   - pip
+  - pulp
   - scikit-learn>=0.22
   - scipy>=1.0
   - spenc
