+++ .github/workflows/unittests.yml
@@ -30,6 +30,9 @@
          with:
            environment-file: ${{ matrix.environment-file }}
            micromamba-version: 'latest'
+       - name: install pulp via pip
+         shell: bash -l {0}
+         run: pip install pulp 
        
        - name: install bleeding edge libpysal (only Ubuntu / Python 3.9)
          shell: bash -l {0}
