+++ environment.yml
@@ -15,4 +15,5 @@ dependencies:
   - pulp
   - scikit-learn>=0.22
   - scipy>=1.0
+  - spaghetti
   - watermark
