### 2021-08-09 fb3b427
Move simulated_geo_points to spopt.locate.util.py
