+++ spopt/locate/__init__.py
@@ -2,3 +2,4 @@ from .base import BaseOutputMixin, CoveragePercentageMixin, MeanDistanceMixin
 from .coverage import LSCP, MCLP
 from .p_median import PMedian
 from .p_center import PCenter
+from .util import simulated_geo_points
