### 2021-08-09 f30b0f2
Improve Test Coverage
- Add tests to facility_client_array method
- Add error tests to FacilityModelBuilder methods
- Add tests to Mixins
