+++ spopt/tests/test_locate.py
@@ -1,3 +1,4 @@
+from spopt.locate.base import FacilityModelBuilder, LocateSolver, T_FacModel
 import numpy
 import geopandas
 import pandas
@@ -9,6 +10,7 @@ from spopt.locate import LSCP, MCLP, PCenter, PMedian
 from spopt.locate.util import simulated_geo_points
 import unittest
 import os
+import pickle
 
 
 class TestGlobalLocate(unittest.TestCase):
@@ -122,38 +124,49 @@ class TestGlobalLocate(unittest.TestCase):
 
 class TestOptimalLocate(unittest.TestCase):
     def setUp(self) -> None:
-        dirpath = os.path.join(os.path.dirname(__file__), "./data/")
+        self.dirpath = os.path.join(os.path.dirname(__file__), "./data/")
         network_distance = pandas.read_csv(
-            dirpath + "SF_network_distance_candidateStore_16_censusTract_205_new.csv"
+            self.dirpath
+            + "SF_network_distance_candidateStore_16_censusTract_205_new.csv"
         )
 
         ntw_dist_piv = network_distance.pivot_table(
             values="distance", index="DestinationName", columns="name"
         )
-        facilities_name = ntw_dist_piv.columns
         demand_name = network_distance[["DestinationName", "demand"]].drop_duplicates()
 
         self.cost_matrix = ntw_dist_piv.to_numpy()
 
         demand_points = pandas.read_csv(
-            dirpath + "SF_demand_205_centroid_uniform_weight.csv"
+            self.dirpath + "SF_demand_205_centroid_uniform_weight.csv"
         )
-        facility_points = pandas.read_csv(dirpath + "SF_store_site_16_longlat.csv")
+        facility_points = pandas.read_csv(self.dirpath + "SF_store_site_16_longlat.csv")
 
-        self.facility_points_gdf = geopandas.GeoDataFrame(
-            facility_points,
-            geometry=geopandas.points_from_xy(
-                facility_points.long, facility_points.lat
-            ),
+        self.facility_points_gdf = (
+            geopandas.GeoDataFrame(
+                facility_points,
+                geometry=geopandas.points_from_xy(
+                    facility_points.long, facility_points.lat
+                ),
+            )
+            .sort_values(by=["NAME"])
+            .reset_index()
         )
-        self.demand_points_gdf = geopandas.GeoDataFrame(
-            demand_points,
-            geometry=geopandas.points_from_xy(demand_points.long, demand_points.lat),
+
+        self.demand_points_gdf = (
+            geopandas.GeoDataFrame(
+                demand_points,
+                geometry=geopandas.points_from_xy(
+                    demand_points.long, demand_points.lat
+                ),
+            )
+            .sort_values(by=["NAME"])
+            .reset_index()
         )
 
         self.service_dist = 5000.0
         self.p_facility = 4
-        self.ai = demand_name.sort_values(by=["DestinationName"])["demand"].to_numpy()
+        self.ai = self.demand_points_gdf["POP2000"].to_numpy()
 
     def test_optimality_lscp_from_cost_matrix(self):
         lscp = LSCP.from_cost_matrix(self.cost_matrix, self.service_dist)
@@ -166,6 +179,16 @@ class TestOptimalLocate(unittest.TestCase):
         lscp = lscp.solve(pulp.PULP_CBC_CMD())
 
         self.assertEqual(lscp.problem.status, pulp.LpStatusInfeasible)
+
+    def test_lscp_facility_client_array_from_cost_matrix(self):
[…]
