### 2021-08-09 e2eb1e8
fix: change assertListEqual to assertCountEqual
