### 2021-08-09 503100e
fix: change assertCountEqual to numpy.testing.assert_array_equal
