### 2021-08-10 0571fe7
fix: test only windows GHA
