### 2021-08-10 0bcd8e0
fix: test only windows python-39 - GHA
