### 2021-08-10 2e78c7f
fix: skipping tests for windows
- mclp_facility_client_array_from_geodataframe
- pcenter_facility_client_array_from_geodataframe
