+++ .github/workflows/unittests.yml
@@ -15,11 +15,11 @@
      timeout-minutes: 30 
      strategy:
        matrix:
-         os: ['windows-latest'] #['macos-latest', 'ubuntu-latest', 'windows-latest']
-         environment-file: [.ci/39.yaml] #[.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
-        #  include:
-        #    - environment-file: .ci/39-DEV.yaml
-        #      os: ubuntu-latest
+         os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
+         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
+         include:
+           - environment-file: .ci/39-DEV.yaml
+             os: ubuntu-latest
     
      steps:
        - name: checkout repo
