### 2021-08-10 362d6d2
fix: mistake with packages in test_locate.py
