### 2021-08-10 8610b37
test: improve test coverage
- Chack ValueError for crs different when creating Locate Models
- Check warning of mixed geometry type in geodataframe
