### 2021-08-10 b3fa600
test: improve test coverage
- Test client_facility_array
- Rename .pkl -> objective to fac2cli
