+++ spopt/tests/test_locate.py
@@ -189,7 +189,7 @@ class TestOptimalLocate(unittest.TestCase):
         self.assertEqual(lscp.problem.status, pulp.LpStatusInfeasible)
 
     def test_lscp_facility_client_array_from_cost_matrix(self):
-        with open(self.dirpath + "lscp_objective.pkl", "rb") as f:
+        with open(self.dirpath + "lscp_fac2cli.pkl", "rb") as f:
             lscp_objective = pickle.load(f)
 
         lscp = LSCP.from_cost_matrix(self.cost_matrix, self.service_dist)
@@ -197,6 +197,17 @@ class TestOptimalLocate(unittest.TestCase):
         lscp.facility_client_array()
 
         numpy.testing.assert_array_equal(lscp.fac2cli, lscp_objective)
+
+    def test_lscp_client_facility_array_from_cost_matrix(self):
+        with open(self.dirpath + "lscp_cli2fac.pkl", "rb") as f:
+            lscp_objective = pickle.load(f)
+
+        lscp = LSCP.from_cost_matrix(self.cost_matrix, self.service_dist)
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+        lscp.facility_client_array()
+        lscp.client_facility_array()
+
+        numpy.testing.assert_array_equal(lscp.cli2fac, lscp_objective)
 
     def test_optimality_lscp_from_geodataframe(self):
         lscp = LSCP.from_geodataframe(
@@ -222,7 +233,7 @@ class TestOptimalLocate(unittest.TestCase):
         self.assertEqual(lscp.problem.status, pulp.LpStatusInfeasible)
 
     def test_lscp_facility_client_array_from_geodataframe(self):
-        with open(self.dirpath + "lscp_geodataframe_objective.pkl", "rb") as f:
+        with open(self.dirpath + "lscp_geodataframe_fac2cli.pkl", "rb") as f:
             lscp_objective = pickle.load(f)
 
         lscp = LSCP.from_geodataframe(
@@ -236,6 +247,23 @@ class TestOptimalLocate(unittest.TestCase):
         lscp.facility_client_array()
 
         numpy.testing.assert_array_equal(lscp.fac2cli, lscp_objective)
+
+    def test_lscp_client_facility_array_from_geodataframe(self):
+        with open(self.dirpath + "lscp_geodataframe_cli2fac.pkl", "rb") as f:
+            lscp_objective = pickle.load(f)
+
+        lscp = LSCP.from_geodataframe(
+            self.demand_points_gdf,
+            self.facility_points_gdf,
+            "geometry",
+            "geometry",
+            self.service_dist,
+        )
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+        lscp.facility_client_array()
+        lscp.client_facility_array()
+
+        numpy.testing.assert_array_equal(lscp.cli2fac, lscp_objective)
 
     def test_optimality_mclp_from_cost_matrix(self):
         mclp = MCLP.from_cost_matrix(
@@ -258,7 +286,7 @@ class TestOptimalLocate(unittest.TestCase):
         self.assertEqual(mclp.problem.status, pulp.LpStatusInfeasible)
 
     def test_mclp_facility_client_array_from_cost_matrix(self):
-        with open(self.dirpath + "mclp_objective.pkl", "rb") as f:
+        with open(self.dirpath + "mclp_fac2cli.pkl", "rb") as f:
             mclp_objective = pickle.load(f)
 
         mclp = MCLP.from_cost_matrix(
@@ -272,8 +300,24 @@ class TestOptimalLocate(unittest.TestCase):
 
         numpy.testing.assert_array_equal(mclp.fac2cli, mclp_objective)
 
[…]
