### 2021-08-11 747f1e5
enh: delete LOC not used
