+++ spopt/locate/coverage.py
@@ -141,8 +141,6 @@ class LSCP(LocateSolver, BaseOutputMixin):
         dem_data = np.array([dem.x.to_numpy(), dem.y.to_numpy()]).T
         fac_data = np.array([fac.x.to_numpy(), fac.y.to_numpy()]).T
 
-        distances = np.array([])
-
         if gdf_demand.crs != gdf_fac.crs:
             raise ValueError(
                 f"geodataframes crs are different: gdf_demand-{gdf_demand.crs}, gdf_fac-{gdf_fac.crs}"
@@ -339,8 +337,6 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
 
         dem_data = np.array([dem.x.to_numpy(), dem.y.to_numpy()]).T
         fac_data = np.array([fac.x.to_numpy(), fac.y.to_numpy()]).T
-
-        distances = np.array([])
 
         if gdf_demand.crs != gdf_fac.crs:
             raise ValueError(
