### 2021-08-12 13b935d
allow for precomputed distance matrices
