### 2021-08-12 3367223
move to minimizing the full attribute kernel rather than the spatially-reduced dissimilarity
