### 2021-08-16 8daf4b9
chore: move tests fac_client and client_fac to synthetic tests
