+++ spopt/tests/test_locate.py
@@ -21,8 +21,10 @@ else:
     WINDOWS = False
 
 
-class TestGlobalLocate(unittest.TestCase):
+class TestSyntheticLocate(unittest.TestCase):
     def setUp(self) -> None:
+        self.dirpath = os.path.join(os.path.dirname(__file__), "./data/")
+
         lattice = spaghetti.regular_lattice((0, 0, 10, 10), 9, exterior=True)
         ntw = spaghetti.Network(in_data=lattice)
         gdf = spaghetti.element_as_gdf(ntw, arcs=True)
@@ -50,7 +52,7 @@ class TestGlobalLocate(unittest.TestCase):
         )
 
         self.facilities_snapped = spaghetti.element_as_gdf(
-            ntw, pp_name="clients", snapped=True
+            ntw, pp_name="facilities", snapped=True
         )
 
         self.cost_matrix = ntw.allneighbordistances(
@@ -60,10 +62,33 @@ class TestGlobalLocate(unittest.TestCase):
 
         self.ai = numpy.random.randint(1, 12, client_count)
 
+        self.clients_snapped["weights"] = self.ai
+
     def test_lscp_from_cost_matrix(self):
         lscp = LSCP.from_cost_matrix(self.cost_matrix, 10)
         result = lscp.solve(pulp.PULP_CBC_CMD())
         self.assertIsInstance(result, LSCP)
+
+    def test_lscp_facility_client_array_from_cost_matrix(self):
+        with open(self.dirpath + "lscp_fac2cli.pkl", "rb") as f:
+            lscp_objective = pickle.load(f)
+
+        lscp = LSCP.from_cost_matrix(self.cost_matrix, 8)
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+        lscp.facility_client_array()
+
+        numpy.testing.assert_array_equal(lscp.fac2cli, lscp_objective)
+
+    def test_lscp_client_facility_array_from_cost_matrix(self):
+        with open(self.dirpath + "lscp_cli2fac.pkl", "rb") as f:
+            lscp_objective = pickle.load(f)
+
+        lscp = LSCP.from_cost_matrix(self.cost_matrix, 8)
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+        lscp.facility_client_array()
+        lscp.client_facility_array()
+
+        numpy.testing.assert_array_equal(lscp.cli2fac, lscp_objective)
 
     def test_lscp_from_geodataframe(self):
         lscp = LSCP.from_geodataframe(
@@ -72,6 +97,39 @@ class TestGlobalLocate(unittest.TestCase):
         result = lscp.solve(pulp.PULP_CBC_CMD())
         self.assertIsInstance(result, LSCP)
 
+    def test_lscp_facility_client_array_from_geodataframe(self):
+        with open(self.dirpath + "lscp_geodataframe_fac2cli.pkl", "rb") as f:
+            lscp_objective = pickle.load(f)
+
+        lscp = LSCP.from_geodataframe(
+            self.clients_snapped,
+            self.facilities_snapped,
+            "geometry",
+            "geometry",
+            8,
+        )
+        lscp = lscp.solve(pulp.PULP_CBC_CMD())
+        lscp.facility_client_array()
+
+        numpy.testing.assert_array_equal(lscp.fac2cli, lscp_objective)
+
+    def test_lscp_client_facility_array_from_geodataframe(self):
+        with open(self.dirpath + "lscp_geodataframe_cli2fac.pkl", "rb") as f:
+            lscp_objective = pickle.load(f)
+
+        lscp = LSCP.from_geodataframe(
+            self.clients_snapped,
+            self.facilities_snapped,
+            "geometry",
+            "geometry",
+            8,
+        )
[…]
