### 2021-08-17 #3330 @tmickleydoyle
👋 Thank you all for the invite

I was reading over the checklist of items but I am not able to interact with the markdown by checking the checklist items.  I will continue to work on the checklist async. 🙇
