### 2021-08-20 #3330 @jGaboardi
@tmickleydoyle Yes, that's the one: https://pysal.org/spopt/
