### 2021-08-23 #3330 @jGaboardi
https://pysal.org/spopt/api.html
