+++ spopt/__init__.py
@@ -1,3 +1,5 @@
-__version__ = "0.1.2"
-
 from . import region
+
+from . import _version
+
+__version__ = _version.get_versions()["version"]
