+++ setup.cfg
@@ -1,0 +1,7 @@
+[versioneer]
+VCS = git
+style = pep440
+versionfile_source = spopt/_version.py
+versionfile_build = spopt/_version.py
+tag_prefix = v
+parentdir_prefix = spopt-
