### 2021-09-15 #3330 @tmickleydoyle
Hey @jGaboardi 👋

I opened an [issue](https://github.com/pysal/spopt/issues/195) in the spopt repo about a challenge I am experiencing when trying to import `MCLP`. Thank you 🙇
