### 2021-09-15 #3330 @jGaboardi
@tmickleydoyle Thank you opening the issue. The reason the import is failing is because the `locate` module is not part of `v0.1.1` release that is under review here. The original JOSS paper submission back in May (I believe) by @xf37 focused on regionalization because that was the only module in `spopt` at the time, with no `locate` functionality (the `locate` module is merged into code base but it is not in a release yet and is not published to PyPI or conda-forge). Since then there has been a `v0.1.2` release and several major occurrences:
* @xf37 accepted a position as Asst. Professor.
* We had an incredibly talented GSoC student (@gegen07) that contributed the `locate` module.
* I accepted a new position.

Due to these things I have some questions (and general musings) about moving forward with this submission:
* Should we create a fresh release and rework the submission to include the `locate` module?
* @xf37, do you still have time to spend addressing the concerns of the reviewers for this submission?
* @gegen07 (+@qszhao & @ljwolf), would you like to contribute to this paper if we rework it to include `locate`?
* Also, I need to confirm with my current employer that I can still be involved with my current affiliation.

Pinging some other PySAL core devs for thoughts & advice on this.
* @sjsrey, @knaaptime, @slumnitz, @TaylorOshan, @jeffcsauer, @martinfleis, @darribas
