### 2021-09-15 pysal/spopt#195 @jGaboardi
The `locate` module has been merged into the code base, but it is not yet part of a release. However, It will be in the next release.
