### 2021-09-16 pysal/spopt#195 @tmickleydoyle
Great! I think I used a code sample from the README file, but I will check again. Thank you for the prompt response.
