### 2021-09-16 #3330 @xf37
@jGaboardi <https://github.com/jGaboardi> Thanks for taking care of this. I
am back to work now.
I am fine with both, either keeping this version or creating a new release.
Just let me know your thoughts @gegen07 <https://github.com/gegen07> (+
@qszhao <https://github.com/qszhao> & @ljwolf <https://github.com/ljwolf>)
@sjsrey <https://github.com/sjsrey>, @knaaptime
<https://github.com/knaaptime>

On Wed, Sep 15, 2021 at 6:45 PM James Gaboardi ***@***.***>
wrote:

> @tmickleydoyle <https://github.com/tmickleydoyle> Thank you opening the
> issue. The reason the import is failing is because the locate module is
> not part of v0.1.1 release that is under review here. The original JOSS
> paper submission back in May (I believe) by @xf37
> <https://github.com/xf37> focused on regionalization because that was the
> only module in spopt at the time, with no locate functionality (the locate
> module is merged into code base but it is not in a release yet and is not
> published to PyPI or conda-forge). Since then there has been a v0.1.2
> release and several major occurrences:
>
>    - @xf37 <https://github.com/xf37> accepted a position as Asst.
>    Professor.
>    - We had an incredibly talented GSoC student ***@***.***
>    <https://github.com/gegen07>) that contributed the locate module.
>    - I accepted a new position.
>
> Due to these things I have some questions (and general musings) about
> moving forward with this submission:
>
>    - Should we create a fresh release and rework the submission to
>    include the locate module?
>    - @xf37 <https://github.com/xf37>, do you still have time to spend
>    addressing the concerns of the reviewers for this submission?
>    - @gegen07 <https://github.com/gegen07> ***@***.***
>    <https://github.com/qszhao> & @ljwolf <https://github.com/ljwolf>),
>    would you like to contribute to this paper if we rework it to include
>    locate?
>    - Also, I need to confirm with my current employer that I can still be
>    involved with my current affiliation.
>
> Pinging some other PySAL core devs for thoughts & advice on this.
>
>    - @sjsrey <https://github.com/sjsrey>, @knaaptime
>    <https://github.com/knaaptime>, @slumnitz <https://github.com/slumnitz>,
>    @TaylorOshan <https://github.com/TaylorOshan>, @jeffcsauer
>    <https://github.com/jeffcsauer>, @martinfleis
>    <https://github.com/martinfleis>, @darribas
>    <https://github.com/darribas>
>
