### 2021-09-17 #3330 @jGaboardi
@arfon @vissarion Do you have any thoughts on this? @xf37 is OK for reworking the manuscript and we would like @gegen07 to get onboard to highlight his achievements in the `locate` module.
