### 2021-09-17 #3330 @vissarion
@jGaboardi if `locate` module is the best way to address the [review issue](https://github.com/pysal/spopt/issues/195) I agree to work on including it (either in this release or in a fresh one) and also update the JOSS paper text.
