### 2021-09-18 pysal/spopt#196 @jGaboardi
rework JOSS manuscript
xref  openjournals/joss-reviews#3330 ([comment](https://github.com/openjournals/joss-reviews/issues/3330#issuecomment-921717815))

We should start thinking about how the revised manuscript will look. @xf37 originally did the draft as a google doc. What do we think about continuing with that? Also, we need to get a release out that includes the `locate` module (prereq #190).

@ljwolf & @qszhao, would you like to get in on the paper, considering your mentorship in GSoC (and Levi's SKATER, SPENC, etc. stuff in `spopt`)?

@gegen07, @sjsrey, @knaaptime, @rwei5
