### 2021-09-18 pysal/spopt#196 @qszhao
@jGaboardi I would be very happy to help with the manuscript if it is necessary...
