### 2021-09-24 a1f02b3
doc: add examples section
