### 2021-09-26 db04d21
doc: add spaghetti as requirements
