### 2021-10-13 pysal/spopt#196 @gegen07
For example section for locate module, should we use the real-world dataset or the example dataset using `spaghetti` as written in [readme](https://github.com/pysal/spopt/blob/main/README.md) is enough?
