### 2021-10-17 06c5e23
adapt release workflow for autonotes
