### 2021-10-17 2f7d95f
add autonotes customizations
