### 2021-10-20 #3330 @vissarion
@jGaboardi how major is the revision / rework you are working on? Does it make sense for a reviewer that didn't start the review yet to start it now or is it better to wait until you finish https://github.com/pysal/spopt/issues/196 ?
