### 2021-10-27 9e65f29
Update candidates list after making a move
