### 2021-10-30 279828a
update paper, bib, & maxp fig
