### 2021-11-03 #3330 @jGaboardi
@vissarion Our apologies for the delays here. We have now closed out pysal/spopt#196 with updated manuscript text and merged `locate` functionality. The fresh release in [`v0.2.1`](https://github.com/pysal/spopt/releases/tag/v0.2.1).
