### 2021-11-03 pysal/spopt#195 @jGaboardi
@tmickleydoyle This is solved with in [`v0.2.1`](https://github.com/pysal/spopt/releases/tag/v0.2.1).
