### 2021-11-19 #3330 @jGaboardi
The acknowledgement that applies to my authorship renders properly as a block quote in [`paper.md`](https://github.com/pysal/spopt/blob/main/paper/paper.md#acknowledgements), but not in the proof. Any advice on how to correct what I'm doing wrong?

@vissarion @arfon
