### 2021-11-19 #3330 @martinfleis
@jGaboardi -> https://github.com/pysal/spopt/pull/207
