+++ setup.py
@@ -56,6 +56,7 @@ def setup_package():
             "Programming Language :: Python :: 3.7",
             "Programming Language :: Python :: 3.8",
             "Programming Language :: Python :: 3.9",
+            "Programming Language :: Python :: 3.10",
         ],
         license="3-Clause BSD",
         packages=find_packages(),
