+++ codecov.yml
@@ -1,6 +1,6 @@
 codecov:
   notify:
-    after_n_builds: 7
+    after_n_builds: 10
 coverage:
   range: 50..95
   round: nearest
@@ -18,5 +18,5 @@ coverage:
 comment:
   layout: "reach, diff, files"
   behavior: once
-  after_n_builds: 7
+  after_n_builds: 10
   require_changes: true
