+++ .ci/310.yaml
@@ -2,15 +2,16 @@ name: test
 channels:
   - conda-forge
 dependencies:
-  - python=3.9
+  - python=3.10
   - geopandas>=0.7.0
+  - libpysal
   - matplotlib
   - networkx
   - numpy>=1.3
   - pandas>=1
-  - pip
   - scikit-learn>=0.22
   - scipy>=1.0
+  - shapely
   - spenc
   - spaghetti
   # testing
@@ -18,3 +19,9 @@ dependencies:
   - coverage
   - pytest
   - pytest-cov
+  # docs
+  - nbsphinx
+  - numpydoc
+  - sphinx
+  - sphinxcontrib-bibtex
+  - sphinx_bootstrap_theme
