+++ .github/workflows/unittests.yml
@@ -16,9 +16,9 @@
      strategy:
        matrix:
          os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
-         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
+         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml, .ci/310.yaml]
          include:
-           - environment-file: .ci/39-DEV.yaml
+           - environment-file: .ci/310-DEV.yaml
              os: ubuntu-latest
     
      steps:
@@ -34,7 +34,7 @@
          shell: bash -l {0}
          run: pip install pulp 
        
-       - name: install bleeding edge libpysal (only Ubuntu / Python 3.9)
+       - name: install bleeding edge libpysal (only Ubuntu / Python 3.10)
          shell: bash -l {0}
          run: pip install git+https://github.com/pysal/libpysal.git@master
          if: matrix.os == 'ubuntu-latest' && contains(matrix.environment-file, 'DEV')
