+++ .ci/310-DEV.yaml
@@ -1,0 +1,20 @@
+name: test
+channels:
+  - conda-forge
+dependencies:
+  - python=3.10
+  - geopandas>=0.7.0
+  - matplotlib
+  - networkx
+  - numpy>=1.3
+  - pandas>=1
+  - pip
+  - scikit-learn>=0.22
+  - scipy>=1.0
+  - spenc
+  - spaghetti
+  # testing
+  - codecov
+  - coverage
+  - pytest
+  - pytest-cov
