### 2021-12-24 a8cce23
slim down testing matrix
