+++ .github/workflows/unittests.yml
@@ -15,11 +15,15 @@
      timeout-minutes: 30 
      strategy:
        matrix:
-         os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
-         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml, .ci/310.yaml]
+         os: [ubuntu-latest]
+         environment-file: [ci/37.yaml, ci/38.yaml, ci/39.yaml, ci/310.yaml]
          include:
-           - environment-file: .ci/310-DEV.yaml
+           - environment-file: ci/310-DEV.yaml
              os: ubuntu-latest
+           - environment-file: ci/310.yaml
+             os: macos-latest
+           - environment-file: ci/310.yaml
+             os: windows-latest
     
      steps:
        - name: checkout repo
