+++ .github/workflows/unittests.yml
@@ -16,13 +16,13 @@
      strategy:
        matrix:
          os: [ubuntu-latest]
-         environment-file: [ci/37.yaml, ci/38.yaml, ci/39.yaml, ci/310.yaml]
+         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml, .ci/310.yaml]
          include:
-           - environment-file: ci/310-DEV.yaml
+           - environment-file: .ci/310-DEV.yaml
              os: ubuntu-latest
-           - environment-file: ci/310.yaml
+           - environment-file: .ci/310.yaml
              os: macos-latest
-           - environment-file: ci/310.yaml
+           - environment-file: .ci/310.yaml
              os: windows-latest
     
      steps:
