### 2021-12-24 8afd43e
streamline matrix logic
