+++ .github/workflows/unittests.yml
@@ -16,10 +16,8 @@
      strategy:
        matrix:
          os: [ubuntu-latest]
-         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml, .ci/310.yaml]
+         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml, .ci/310.yaml, .ci/310-DEV.yaml]
          include:
-           - environment-file: .ci/310-DEV.yaml
-             os: ubuntu-latest
            - environment-file: .ci/310.yaml
              os: macos-latest
            - environment-file: .ci/310.yaml
