### 2021-12-24 1abfc56
add 310 to CI matrix (#208)
* add 310 to CI matrix

* resolve collections.Iterable failure

* slim down testing matrix

* path correction

* streamline matrix logic
