+++ .github/workflows/unittests.yml
@@ -15,11 +15,13 @@
      timeout-minutes: 30 
      strategy:
        matrix:
-         os: ['macos-latest', 'ubuntu-latest', 'windows-latest']
-         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml]
+         os: [ubuntu-latest]
+         environment-file: [.ci/37.yaml, .ci/38.yaml, .ci/39.yaml, .ci/310.yaml, .ci/310-DEV.yaml]
          include:
-           - environment-file: .ci/39-DEV.yaml
-             os: ubuntu-latest
+           - environment-file: .ci/310.yaml
+             os: macos-latest
+           - environment-file: .ci/310.yaml
+             os: windows-latest
     
      steps:
        - name: checkout repo
@@ -34,7 +36,7 @@
          shell: bash -l {0}
          run: pip install pulp 
        
-       - name: install bleeding edge libpysal (only Ubuntu / Python 3.9)
+       - name: install bleeding edge libpysal (only Ubuntu / Python 3.10)
          shell: bash -l {0}
          run: pip install git+https://github.com/pysal/libpysal.git@master
          if: matrix.os == 'ubuntu-latest' && contains(matrix.environment-file, 'DEV')
