+++ .ci/37.yaml
@@ -19,3 +19,4 @@ dependencies:
   - coverage
   - pytest
   - pytest-cov
+  - pytest-xdist
