+++ .ci/310.yaml
@@ -19,6 +19,7 @@ dependencies:
   - coverage
   - pytest
   - pytest-cov
+  - pytest-xdist
   # docs
   - nbsphinx
   - numpydoc
