### 2022-01-17 4a82dfa
Xdist tests (#211)
* add pytest-xdist to reqs

* update pytest command call
