+++ .github/workflows/unittests.yml
@@ -12,7 +12,9 @@
    unittests:
      name: ${{ matrix.os }}, ${{ matrix.environment-file }}
      runs-on: ${{ matrix.os }}
-     timeout-minutes: 30 
+     timeout-minutes: 30
+     env:
+       RUN_TEST: pytest spopt --cov spopt -v -n auto --cov-config .coveragerc --cov-report xml --color yes --cov-append --cov-report term-missing
      strategy:
        matrix:
          os: [ubuntu-latest]
@@ -43,12 +45,12 @@
        
        - name: run tests - bash
          shell: bash -l {0}
-         run: pytest -v spopt --cov=spopt --cov-report=xml
+         run: ${{ env.RUN_TEST }}
          if: matrix.os != 'windows-latest'
        
        - name: run tests - powershell
          shell: powershell
-         run: pytest -v spopt --cov=spopt --cov-report=xml
+         run: ${{ env.RUN_TEST }}
          if: matrix.os == 'windows-latest'
        
        - name: codecov
