+++ .ci/310-DEV.yaml
@@ -18,3 +18,4 @@ dependencies:
   - coverage
   - pytest
   - pytest-cov
+  - pytest-xdist
