+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos: 
 -   repo: https://github.com/ambv/black
-    rev: 21.6b0
+    rev: 21.12b0
     hooks: 
     - id: black
       language_version: python3.9
