### 2022-02-04 c5eae2c
docs: instance has attributes not returns
