### 2022-02-04 3dc23ec
docs: add author info and update k-means
