+++ spopt/region/maxp.py
@@ -6,6 +6,9 @@ regionalization for spatially explicit neighborhood delineation." International
 Journal of Geographical Information Science. Accepted 2020-04-12.
 """
 
+__author__ = ["Ran Wei", "Serge Rey", "Elijah Knaap"]
+__email__ = "sjsrey@gmail.com"
+
 from ..BaseClass import BaseSpOptHeuristicSolver
 from .base import (
     w_to_g,
