### 2022-02-04 0b289ba
docs: localize azp docstrings
