### 2022-02-06 #3330 @jGaboardi
> > > @vissarion @arfon We are wondering if there is anything we can do to help move this along?
> >
> >
> > I have pinged @ArikaLZ by email, let's wait a few days for her reply.
>
> Unfortunately, I have been notified by @ArikaLZ that they cannot review this paper due to time constraints. I am in a quest of a new reviewer.
>
> @jGaboardi if you have any suggestions please let me know.

@vissarion All my suggestions are tapped out. LOL.
