+++ .github/workflows/release_and_publish.yml
@@ -29,7 +29,7 @@ jobs:
         python setup.py sdist bdist_wheel
     
     - name: Create Release Notes
-      uses: actions/github-script@v5
+      uses: actions/github-script@v6
       with:
         github-token: ${{secrets.GITHUB_TOKEN}}
         script: |
