### 2022-02-28 3ecc223
build(deps): bump actions/setup-python from 2 to 3.0.0
Bumps [actions/setup-python](https://github.com/actions/setup-python) from 2 to 3.0.0.
- [Release notes](https://github.com/actions/setup-python/releases)
- [Commits](https://github.com/actions/setup-python/compare/v2...v3.0.0)

---
updated-dependencies:
- dependency-name: actions/setup-python
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
