+++ .github/workflows/release_and_publish.yml
@@ -18,7 +18,7 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python
-      uses: actions/setup-python@v3.0.0
+      uses: actions/setup-python@v3
       with:
         python-version: '3.x'
     
