+++ .github/workflows/build_docs.yml
@@ -27,7 +27,7 @@
      
      steps:
        - name: checkout repo
-         uses: actions/checkout@v2
+         uses: actions/checkout@v3
        
        - name: setup micromamba
          uses: mamba-org/provision-with-micromamba@main
