### 2022-03-13 ce26fa8
feat: policy for multiple components in maxp
