+++ spopt/region/maxp.py
@@ -22,13 +22,136 @@ from .base import (
 )
 
 from scipy.spatial.distance import pdist, squareform
+from scipy.spatial import KDTree
+from scipy.sparse.csgraph import connected_components
+import libpysal
 import geopandas as gp
 import numpy as np
 from copy import deepcopy
-from scipy.sparse.csgraph import connected_components
 
 ITERCONSTRUCT = 999
 ITERSA = 10
+
+
+def infeasible_components(gdf, w, threshold_var, threshold):
+    """Identify infeasible components
+    
+    Parameters
+    ---------- 
+
+    gdf : geopandas.GeoDataFrame, required
+        Geodataframe containing original data
+
+    w : libpysal.weights.W, required
+        Weights object created from given data
+
+    attrs_name : list, required
+        Strings for attribute names to measure similarity (cols of ``geopandas.GeoDataFrame``).
+
+    threshold_var : string, requied
+        The name of the spatial extensive attribute variable.
+
+    threshold : {int, float}, required
+        The threshold value.
+
+    Returns
+    -------
+    list of infeasible components
+    """
+    components = np.unique(w.component_labels)
+    if len(components) > 1:
+        gdf['_components'] = w.component_labels
+        gb = gdf.groupby(by='_components').sum()
+        gdf.drop(columns="_components", inplace=True)
+        if gb[threshold_var].min() < threshold:
+            l = gb[gb[threshold_var]< threshold]
+            return l.index.values.tolist()
+    return []
+
+
+def plot_components(gdf, w):
+    """Helper plot to view components of the W for a gdf
+    
+    Parameters
+    ----------
+    
+    gdf: geopandas.GeoDataframe
+
+    w: libpysal.weights.W defined on gdf
+    
+    Returns
+    -------
+    folium.folium.Map
+
+    """
+    cgdf = gdf.copy()
+    cgdf['component'] = w.component_labels
+    return cgdf.explore(column='component', categorical=True)
+
+
+def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
+    """Modify infeasible components
+    
+    Parameters
+    ---------- 
+
+    gdf : geopandas.GeoDataFrame, required
+        Geodataframe containing original data
+
+    w : libpysal.weights.W, required
+        Weights object created from given data
+
+    attrs_name : list, required
+        Strings for attribute names to measure similarity (cols of ``geopandas.GeoDataFrame``).
+
+    threshold_var : string, requied
+        The name of the spatial extensive attribute variable.
+
+    threshold : {int, float}, required
+        The threshold value.
+    
+    policy: str
+          'attach' will attach areas of infeasible components to
+          nearest neighbor in a feasible component.
+          'keep' keeps infeasible components and attempts to solve.
+
+
+    Returns
+    -------
+    gdf: geopandas.GeoDataFrame
+
+ 
+    w : libpysal.weights.W, required
+        Weights object created from given data
+    """
+
+    ifcs = infeasible_components(gdf, w, threshold_var, threshold)
+
+    if ifcs == np.unique(w.component_labels).tolist():
+        print('No feasible components.')
+        return gdf, w
+    policy = policy.lower()
+    if not ifcs or policy == 'keep':
+        return gdf, w
+    elif policy == 'attach':
+        fcs = [c for c in w.component_labels if c not in ifcs]
+        ifcas = np.where(np.isin(w.component_labels, ifcs))[0]
+        fcas = np.where(~np.isin(w.component_labels, ifcs))[0]
+        tree = KDTree(list(zip(gdf.iloc[fcas].geometry.centroid.x,
+                               gdf.iloc[fcas].geometry.centroid.y)))
[…]
