### 2022-03-14 01c1610
feat: add drop as a component policy
