### 2022-03-14 086476d
doc: component policy
