### 2022-03-14 a1a9f42
feat: handle case with no feasible components
