+++ spopt/region/maxp.py
@@ -10,22 +10,11 @@ __author__ = ["Ran Wei", "Serge Rey", "Elijah Knaap"]
 __email__ = "sjsrey@gmail.com"
 
 from ..BaseClass import BaseSpOptHeuristicSolver
-from .base import (
-    w_to_g,
-    move_ok,
-    ok_moves,
-    region_neighbors,
-    _centroid,
-    _closest,
-    _seeds,
-    is_neighbor,
-)
 
 from scipy.spatial.distance import pdist, squareform
 from scipy.spatial import KDTree
 from scipy.sparse.csgraph import connected_components
 import libpysal
-import geopandas as gp
 import numpy as np
 from copy import deepcopy
 
@@ -35,9 +24,9 @@ ITERSA = 10
 
 def infeasible_components(gdf, w, threshold_var, threshold):
     """Identify infeasible components
-    
+
     Parameters
-    ---------- 
+    ----------
 
     gdf : geopandas.GeoDataFrame, required
         Geodataframe containing original data
@@ -46,7 +35,8 @@ def infeasible_components(gdf, w, threshold_var, threshold):
         Weights object created from given data
 
     attrs_name : list, required
-        Strings for attribute names to measure similarity (cols of ``geopandas.GeoDataFrame``).
+        Strings for attribute names to measure similarity
+        (cols of ``geopandas.GeoDataFrame``).
 
     threshold_var : string, requied
         The name of the spatial extensive attribute variable.
@@ -71,14 +61,14 @@ def infeasible_components(gdf, w, threshold_var, threshold):
 
 def plot_components(gdf, w):
     """Helper plot to view components of the W for a gdf
-    
+
     Parameters
     ----------
-    
+
     gdf: geopandas.GeoDataframe
 
     w: libpysal.weights.W defined on gdf
-    
+
     Returns
     -------
     folium.folium.Map
@@ -91,9 +81,9 @@ def plot_components(gdf, w):
 
 def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
     """Modify infeasible components
-    
+
     Parameters
-    ---------- 
+    ----------
 
     gdf : geopandas.GeoDataFrame, required
         Geodataframe containing original data
@@ -109,7 +99,7 @@ def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
 
     threshold : {int, float}, required
         The threshold value.
-    
+
     policy: str
           'attach' will attach areas of infeasible components to
           nearest neighbor in a feasible component.
@@ -120,7 +110,6 @@ def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
     -------
     gdf: geopandas.GeoDataFrame
 
- 
     w : libpysal.weights.W, required
         Weights object created from given data
     """
@@ -128,13 +117,11 @@ def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
     ifcs = infeasible_components(gdf, w, threshold_var, threshold)
 
     if ifcs == np.unique(w.component_labels).tolist():
-        print('No feasible components.')
-        return gdf, w
+        raise Exception("No feasible components found in input.")
     policy = policy.lower()
     if not ifcs or policy == 'keep':
         return gdf, w
     elif policy == 'attach':
-        fcs = [c for c in w.component_labels if c not in ifcs]
         ifcas = np.where(np.isin(w.component_labels, ifcs))[0]
         fcas = np.where(~np.isin(w.component_labels, ifcs))[0]
         tree = KDTree(list(zip(gdf.iloc[fcas].geometry.centroid.x,
[…]
