### 2022-03-15 6ddc509
test: component policy testing
