+++ spopt/tests/test_maxp.py
@@ -2,8 +2,11 @@ import geopandas
 import libpysal
 import numpy
 import unittest
-
+from shapely.geometry import Polygon, box
 from spopt.region import MaxPHeuristic
+from spopt.region.maxp import (infeasible_components,
+                               modify_components,
+                               plot_components)
 
 
 # Mexican states
@@ -15,7 +18,8 @@ class TestMaxPHeuristic(unittest.TestCase):
     def setUp(self):
 
         self.mexico = MEXICO.copy()
-
+        self.mexico['count']=1
+        self.w = libpysal.weights.Queen.from_dataframe(self.mexico)
         # labels for non-verbose and verbose:
         # count=1, threshold=4, top_n=2
         self.basic_labels = [5, 5, 3, 6, 6, 6, 1, 1, 1, 1, 8, 6, 8, 2, 2, 8]
@@ -31,58 +35,92 @@ class TestMaxPHeuristic(unittest.TestCase):
         self.var1_labels = [3, 3, 6, 2, 2, 2, 2, 4, 2, 4, 5, 2, 5, 1, 1, 5, 1]
         self.var1_labels += [4, 5, 5, 1, 1, 3, 3, 6, 3, 6, 4, 4, 6, 6, 4]
 
+        # components
+        n_cols = 5
+        n_rows = 10
+        b = 0
+        h = w = 10
+        component_0 = [box(l*w, b, l*w+w, b+h) for l in range(n_cols)] 
+        b = b + h*2
+        component_1 = [box(l*w, b + h * r, l * w + w, b +  h+ h * r) for r in range(n_rows) for l in range(n_cols) ] 
+        geometries = component_0 + component_1
+
+        self.gdf = geopandas.GeoDataFrame(geometry=geometries, 
+                             data = numpy.ones((n_cols*n_rows+n_cols,1), int),
+                             columns=['var'])
+
     def test_maxp_heuristic_basic_nonverbose(self):
-        self.mexico["count"] = 1
-        w = libpysal.weights.Queen.from_dataframe(self.mexico)
         attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
         threshold = 4
         top_n = 2
         threshold_name = "count"
         numpy.random.seed(123456)
-        args = (self.mexico, w, attrs_name, threshold_name, threshold, top_n)
+        args = (self.mexico, self.w, attrs_name, threshold_name, threshold, top_n)
         model = MaxPHeuristic(*args)
         model.solve()
 
         numpy.testing.assert_array_equal(model.labels_, self.basic_labels)
 
     def test_maxp_heuristic_basic_verbose(self):
-        self.mexico["count"] = 1
-        w = libpysal.weights.Queen.from_dataframe(self.mexico)
         attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
         threshold = 4
         top_n = 2
         threshold_name = "count"
         numpy.random.seed(123456)
-        args = (self.mexico, w, attrs_name, threshold_name, threshold, top_n)
+        args = (self.mexico, self.w, attrs_name, threshold_name, threshold, top_n)
         model = MaxPHeuristic(*args, verbose=True)
         model.solve()
 
         numpy.testing.assert_array_equal(model.labels_, self.basic_labels)
 
     def test_maxp_heuristic_complex(self):
-        self.mexico["count"] = 4
-        w = libpysal.weights.Queen.from_dataframe(self.mexico)
         attrs_name = [f"PCGDP{year}" for year in range(1950, 2010, 10)]
         threshold = 10
         top_n = 5
         threshold_name = "count"
         numpy.random.seed(1999)
[…]
