+++ spopt/region/maxp.py
@@ -1,5 +1,4 @@
-"""
-Max-p regions algorithm
+"""Max-p regions algorithm.
 
 Source: Wei, Ran, Sergio J. Rey, and Elijah Knaap (2020) "Efficient
 regionalization for spatially explicit neighborhood delineation." International
@@ -23,11 +22,10 @@ ITERSA = 10
 
 
 def infeasible_components(gdf, w, threshold_var, threshold):
-    """Identify infeasible components
+    """Identify infeasible components.
 
     Parameters
     ----------
-
     gdf : geopandas.GeoDataFrame, required
         Geodataframe containing original data
 
@@ -49,22 +47,20 @@ def infeasible_components(gdf, w, threshold_var, threshold):
     list of infeasible components
     """
     components = np.unique(w.component_labels)
-    if len(components) > 1:
-        gdf['_components'] = w.component_labels
-        gb = gdf.groupby(by='_components').sum()
-        gdf.drop(columns="_components", inplace=True)
-        if gb[threshold_var].min() < threshold:
-            l = gb[gb[threshold_var]< threshold]
-            return l.index.values.tolist()
+    gdf['_components'] = w.component_labels
+    gb = gdf.groupby(by='_components').sum()
+    gdf.drop(columns="_components", inplace=True)
+    if gb[threshold_var].min() < threshold:
+        l = gb[gb[threshold_var]< threshold]
+        return l.index.values.tolist()
     return []
 
 
 def plot_components(gdf, w):
-    """Helper plot to view components of the W for a gdf
+    """Plot to view components of the W for a gdf.
 
     Parameters
     ----------
-
     gdf: geopandas.GeoDataframe
 
     w: libpysal.weights.W defined on gdf
@@ -80,11 +76,10 @@ def plot_components(gdf, w):
 
 
 def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
-    """Modify infeasible components
+    """Modify infeasible components.
 
     Parameters
     ----------
-
     gdf : geopandas.GeoDataFrame, required
         Geodataframe containing original data
 
@@ -92,7 +87,8 @@ def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
         Weights object created from given data
 
     attrs_name : list, required
-        Strings for attribute names to measure similarity (cols of ``geopandas.GeoDataFrame``).
+        Strings for attribute names to measure similarity (cols of
+        ``geopandas.GeoDataFrame``).
 
     threshold_var : string, requied
         The name of the spatial extensive attribute variable.
@@ -113,7 +109,6 @@ def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
     w : libpysal.weights.W, required
         Weights object created from given data
     """
-
     ifcs = infeasible_components(gdf, w, threshold_var, threshold)
 
     if ifcs == np.unique(w.component_labels).tolist():
@@ -171,7 +166,6 @@ def maxp(
 
     Parameters
     ----------
-
     gdf : geopandas.GeoDataFrame, required
         Geodataframe containing original data
 
@@ -208,7 +202,6 @@ def maxp(
 
     Returns
     -------
-
     max_p : int
         The number of regions.
 
@@ -274,61 +267,11 @@ def maxp(
             if totalWithinRegionDistance < best_obj_value:
                 best_obj_value = totalWithinRegionDistance
                 best_label = finalLabel
-                best_fn = irl
     if verbose:
         print("best objective value:")
         print(best_obj_value)
 
     return max_p, best_label
-
-
-def components_check(
-    arr,
-    attr,
-    threshold_array,
-    distance_matrix,
-    weight,
-    spatialThre,
-    random_assign_choice,
-    max_it=999,
-):
-    """Check for multiple components.
-
-    Parameters
[…]
