### 2022-03-15 8a77cf0
test: update comparisons for new setup in tests
