### 2022-03-15 7150325
refactor: move component checking into base
