+++ spopt/region/base.py
@@ -1,8 +1,10 @@
 """Base classes for spopt/region"""
 
 from libpysal.io.fileio import FileIO as psopen
+import libpysal
 import numpy
 import networkx
+from scipy.spatial import KDTree
 
 
 class RegionMixin(object):
@@ -272,3 +274,132 @@ def is_neighbor(area, region, w):
             neighboring = True
             return neighboring
     return neighboring
+
+
+
+
+def infeasible_components(gdf, w, threshold_var, threshold):
+    """Identify infeasible components.
+
+    Parameters
+    ----------
+    gdf : geopandas.GeoDataFrame, required
+        Geodataframe containing original data
+
+    w : libpysal.weights.W, required
+        Weights object created from given data
+
+    attrs_name : list, required
+        Strings for attribute names to measure similarity
+        (cols of ``geopandas.GeoDataFrame``).
+
+    threshold_var : string, requied
+        The name of the spatial extensive attribute variable.
+
+    threshold : {int, float}, required
+        The threshold value.
+
+    Returns
+    -------
+    list of infeasible components
+    """
+    gdf['_components'] = w.component_labels
+    gb = gdf.groupby(by='_components').sum()
+    gdf.drop(columns="_components", inplace=True)
+    if gb[threshold_var].min() < threshold:
+        l = gb[gb[threshold_var]< threshold]
+        return l.index.values.tolist()
+    return []
+
+
+def plot_components(gdf, w):
+    """Plot to view components of the W for a gdf.
+
+    Parameters
+    ----------
+    gdf: geopandas.GeoDataframe
+
+    w: libpysal.weights.W defined on gdf
+
+    Returns
+    -------
+    folium.folium.Map
+
+    """
+    cgdf = gdf.copy()
+    cgdf['component'] = w.component_labels
+    return cgdf.explore(column='component', categorical=True)
+
+
+def modify_components(gdf, w, threshold_var, threshold, policy='attach'):
+    """Modify infeasible components.
+
+    Parameters
+    ----------
+    gdf : geopandas.GeoDataFrame, required
+        Geodataframe containing original data
+
+    w : libpysal.weights.W, required
+        Weights object created from given data
+
+    attrs_name : list, required
+        Strings for attribute names to measure similarity (cols of
+        ``geopandas.GeoDataFrame``).
+
+    threshold_var : string, requied
+        The name of the spatial extensive attribute variable.
+
+    threshold : {int, float}, required
+        The threshold value.
+
+    policy: str
+          'attach' will attach areas of infeasible components to
+          nearest neighbor in a feasible component.
+          'keep' keeps infeasible components and attempts to solve.
+
+
+    Returns
+    -------
+    gdf: geopandas.GeoDataFrame
+
+    w : libpysal.weights.W, required
+        Weights object created from given data
+    """
+    ifcs = infeasible_components(gdf, w, threshold_var, threshold)
+
+    if ifcs == numpy.unique(w.component_labels).tolist():
+        raise Exception("No feasible components found in input.")
+    policy = policy.lower()
+    if not ifcs or policy == 'keep':
+        return gdf, w
+    elif policy == 'attach':
+        ifcas = numpy.where(numpy.isin(w.component_labels, ifcs))[0]
+        fcas = numpy.where(~numpy.isin(w.component_labels, ifcs))[0]
+        tree = KDTree(list(zip(gdf.iloc[fcas].geometry.centroid.x,
+                               gdf.iloc[fcas].geometry.centroid.y)))
+        query_pnts = list(zip(gdf.iloc[ifcas].geometry.centroid.x,
[…]
