+++ spopt/region/maxp.py
@@ -33,7 +33,7 @@ def maxp(
     max_iterations_construction=ITERCONSTRUCT,
     max_iterations_sa=ITERSA,
     verbose=False,
-    policy='attach',
+    policy='single',
 ):
     """The max-p-regions involves the aggregation of n areas into an unknown maximum number of
     homogeneous regions, while ensuring that each region is contiguous and satisfies a minimum
@@ -69,11 +69,14 @@ def maxp(
         Set to ``True`` for reporting solution progress/debugging.
         Default is ``False``.
     policy : str
-        Defaults to ``attach`` to attach areas from infeasible
-        components to nearest area in a feasible component. ``keep``
-        attempts to solve without modification (useful for
-        debugging). ``drop`` removes areas in infeasible components
-        before solving.
+        Defaults to ``single`` to attach infeasible components using a
+        single linkage between the area in the infeasible component
+        with the smallest nearest neighbor distance to an area in a
+        feasible component. ``multiple`` adds joins for each area
+        in an infeasible component and their nearest neighbor area in a
+        feasible component. ``keep`` attempts to solve without
+        modification (useful for debugging). ``drop`` removes areas in
+        infeasible components before solving.
 
     Returns
     -------
