### 2022-03-15 0f59d3f
test: update tests for new policy options
