### 2022-03-16 3b91ef2
doc: update policy options for maxp
