### 2022-03-16 843e2bc
fix: wrong largest component index
