### 2022-03-18 4495a6c
doc: update docs for component_policy
