+++ .github/workflows/unittests.yml
@@ -54,7 +54,7 @@
          if: matrix.os == 'windows-latest'
        
        - name: codecov
-         uses: codecov/codecov-action@v2
+         uses: codecov/codecov-action@v3
          with:
            token: ${{ secrets.CODECOV_TOKEN }}
            file: ./coverage.xml
