### 2022-04-17 72059b2
chore: add predefined locations to MCLP
