+++ spopt/locate/coverage.py
@@ -316,6 +316,7 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
         weights: np.array,
         max_coverage: float,
         p_facilities: int,
+        predefined_facilities_arr: np.array = None,
         name: str = "MCLP",
     ):
         """
@@ -331,6 +332,8 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
             maximum acceptable service distance by problem
         p_facilities: int
             number of facilities to be located
+        predefined_facilities_arr: np.array, default=None
+            bool array defining which facilities are already defined
         name: str, default="MCLP"
             name of the problem
 
@@ -402,10 +405,16 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
         weights = np.reshape(weights, (cost_matrix.shape[0], 1))
 
         mclp.__add_obj(weights, r_cli)
+        
+        if predefined_facilities_arr is not None:
+            FacilityModelBuilder.add_predefined_facility_constraint(mclp, mclp.problem, predefined_facilities_arr)
+        
         FacilityModelBuilder.add_maximal_coverage_constraint(
             mclp, mclp.problem, mclp.aij, r_fac, r_cli
         )
+        
         FacilityModelBuilder.add_facility_constraint(mclp, mclp.problem, p_facilities)
+
 
         return mclp
 
@@ -419,6 +428,7 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
         weights_cols: str,
         max_coverage: float,
         p_facilities: int,
+        predefined_facility_col: str = None,
         distance_metric: str = "euclidean",
         name: str = "MCLP",
     ):
@@ -442,6 +452,8 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
             maximum acceptable service distance by problem
         p_facilities: int
             number of facilities to be located
+        predefined_facility_col: np.array, default=None
+            column name representing facilities are already defined
         distance_metric: str, default="euclidean"
             metrics supported by :method: `scipy.spatial.distance.cdist` used for the distance calculations
         name: str, default="MCLP"
@@ -505,6 +517,11 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
         >>> mclp_from_geodataframe.facility_client_array()
         >>> mclp_from_geodataframe.fac2cli
         """
+
+        predefined_facilities_arr = None
+        if predefined_facility_col is not None:
+            predefined_facilities_arr = gdf_fac[predefined_facility_col].to_numpy()
+
         service_load = gdf_demand[weights_cols].to_numpy()
         dem = gdf_demand[demand_col]
         fac = gdf_fac[facility_col]
@@ -537,7 +554,7 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
         distances = cdist(dem_data, fac_data, distance_metric)
 
         return cls.from_cost_matrix(
-            distances, service_load, max_coverage, p_facilities, name
+            distances, service_load, max_coverage, p_facilities, predefined_facilities_arr, name
         )
 
     def facility_client_array(self) -> None:
