### 2022-04-19 f7a7c1e
chore: add predefined locations to LSCP
