+++ spopt/locate/coverage.py
@@ -44,7 +44,7 @@ class LSCP(LocateSolver, BaseOutputMixin):
 
     @classmethod
     def from_cost_matrix(
-        cls, cost_matrix: np.array, max_coverage: float, name: str = "LSCP"
+        cls, cost_matrix: np.array, max_coverage: float, predefined_facilities_arr: np.array = None, name: str = "LSCP"
     ):
         """
         Create a LSCP object based on cost matrix.
@@ -119,6 +119,9 @@ class LSCP(LocateSolver, BaseOutputMixin):
         lscp.aij = np.zeros(cost_matrix.shape)
         lscp.aij[cost_matrix <= max_coverage] = 1
 
+        if predefined_facilities_arr is not None:
+            FacilityModelBuilder.add_predefined_facility_constraint(lscp, lscp.problem, predefined_facilities_arr)
+
         lscp.__add_obj()
         FacilityModelBuilder.add_set_covering_constraint(
             lscp, lscp.problem, lscp.aij, r_fac, r_cli
@@ -134,6 +137,7 @@ class LSCP(LocateSolver, BaseOutputMixin):
         demand_col: str,
         facility_col: str,
         max_coverage: float,
+        predefined_facility_col: str = None,
         distance_metric: str = "euclidean",
         name: str = "LSCP",
     ):
@@ -205,6 +209,10 @@ class LSCP(LocateSolver, BaseOutputMixin):
 
         """
 
+        predefined_facilities_arr = None
+        if predefined_facility_col is not None:
+            predefined_facilities_arr = gdf_fac[predefined_facility_col].to_numpy()
+
         dem = gdf_demand[demand_col]
         fac = gdf_fac[facility_col]
 
@@ -235,7 +243,7 @@ class LSCP(LocateSolver, BaseOutputMixin):
 
         distances = cdist(dem_data, fac_data, distance_metric)
 
-        return cls.from_cost_matrix(distances, max_coverage, name)
+        return cls.from_cost_matrix(distances, max_coverage, predefined_facilities_arr, name)
 
     def facility_client_array(self) -> None:
         """
