### 2022-05-18 #3330 @vissarion
Hi, @apulverizer, @samuel-rosa, @michstrauch, @ncxiao, @rowanc1, @sarats, @MichaelHoltonPrice, @PaulDebus.

Would any of you be willing to review this submission for JOSS?
