### 2022-05-30 #3330 @ArikaLZ
Dear Vissarion,
I ended up in a hospital a month ago and everything is now in hold. I sincerely apologize for not sticking to the deadlines and I hope to review papers for JOSS in the future.
Apilogies again and best wishes,
Arika

Get Outlook for Android<https://aka.ms/AAb9ysg>
