### 2022-05-31 #3330 @vissarion
@apulverizer I added you as a reviewer!

JOSS reviews involve downloading and installing the software, and inspecting the repository
and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Please read also about the guidelines https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and the checklist https://joss.readthedocs.io/en/latest/review_checklist.html
