### 2022-06-01 81717ab
check_status method in base locate class
