### 2022-06-01 75a8e56
check solution status following solve
