+++ spopt/tests/test_locate.py
@@ -131,12 +131,14 @@ class TestSyntheticLocate(unittest.TestCase):
         numpy.testing.assert_array_equal(lscp.cli2fac, lscp_objective)
 
     def test_lscp_preselected_facility_client_array_from_geodataframe(self):
-        with open(self.dirpath + "lscp_preselected_loc_geodataframe_fac2cli.pkl", "rb") as f:
+        with open(
+            self.dirpath + "lscp_preselected_loc_geodataframe_fac2cli.pkl", "rb"
+        ) as f:
             lscp_objective = pickle.load(f)
 
         fac_snapped = self.facilities_snapped.copy()
 
-        fac_snapped['predefined_loc'] = numpy.array([0, 0, 0, 0, 1])
+        fac_snapped["predefined_loc"] = numpy.array([0, 0, 0, 0, 1])
 
         lscp = LSCP.from_geodataframe(
             self.clients_snapped,
@@ -144,7 +146,7 @@ class TestSyntheticLocate(unittest.TestCase):
             "geometry",
             "geometry",
             predefined_facility_col="predefined_loc",
-            max_coverage=8
+            max_coverage=8,
         )
         lscp = lscp.solve(pulp.PULP_CBC_CMD(msg=False, warmStart=True))
         lscp.facility_client_array()
@@ -220,14 +222,16 @@ class TestSyntheticLocate(unittest.TestCase):
         mclp.facility_client_array()
 
         numpy.testing.assert_array_equal(mclp.fac2cli, mclp_objective)
-    
+
     def test_mclp_preselected_facility_client_array_from_geodataframe(self):
-        with open(self.dirpath + "mclp_preselected_loc_geodataframe_fac2cli.pkl", "rb") as f:
+        with open(
+            self.dirpath + "mclp_preselected_loc_geodataframe_fac2cli.pkl", "rb"
+        ) as f:
             mclp_objective = pickle.load(f)
 
         fac_snapped = self.facilities_snapped.copy()
 
-        fac_snapped['predefined_loc'] = numpy.array([1, 1, 0, 1, 0])
+        fac_snapped["predefined_loc"] = numpy.array([1, 1, 0, 1, 0])
 
         mclp = MCLP.from_geodataframe(
             self.clients_snapped,
@@ -460,9 +464,8 @@ class TestRealWorldLocate(unittest.TestCase):
 
     def test_infeasibility_lscp_from_cost_matrix(self):
         lscp = LSCP.from_cost_matrix(self.cost_matrix, 20)
-        lscp = lscp.solve(pulp.PULP_CBC_CMD(msg=False))
-
-        self.assertEqual(lscp.problem.status, pulp.LpStatusInfeasible)
+        with self.assertRaises(RuntimeError):
+            lscp.solve(pulp.PULP_CBC_CMD(msg=False))
 
     def test_optimality_lscp_from_geodataframe(self):
         lscp = LSCP.from_geodataframe(
@@ -473,7 +476,6 @@ class TestRealWorldLocate(unittest.TestCase):
             self.service_dist,
         )
         lscp = lscp.solve(pulp.PULP_CBC_CMD(msg=False))
-
         self.assertEqual(lscp.problem.status, pulp.LpStatusOptimal)
 
     def test_infeasibility_lscp_from_geodataframe(self):
@@ -484,8 +486,8 @@ class TestRealWorldLocate(unittest.TestCase):
             "geometry",
             0,
         )
-        lscp = lscp.solve(pulp.PULP_CBC_CMD(msg=False))
-        self.assertEqual(lscp.problem.status, pulp.LpStatusInfeasible)
[…]
