### 2022-06-02 7bcb353
chore: change max_coverage to service_radius
