### 2022-06-04 #3330 @apulverizer
@vissarion I don't appear to have edit access to update the issue checklist. The link to https://github.com/openjournals/joss-reviews/invitations 404s for me. Can you send an invite or change my access level?
