### 2022-06-05 pysal/spopt#240 @apulverizer
Consider documenting class properties
While I was reviewing this package as part of https://github.com/openjournals/joss-reviews/issues/3330 I noticed that quite a few of the tutorials use undocumented properties of the models for plotting and/or statistics.

For example:
https://github.com/pysal/spopt/blob/360ea6cbfdd81655874811f50dc26db865ad533f/spopt/locate/coverage.py#L277

is used in the sample MCLP notebook: https://github.com/pysal/spopt/blob/main/notebooks/mclp.ipynb?short_path=d3109c7#L944

I couldn't find any doc on this property in the [API ref](https://pysal.org/spopt/generated/spopt.locate.coverage.MCLP.html#spopt.locate.coverage.MCLP). From a coding perspective I expected `facility_client_array` to return the 2x2 matrix, not `None`. It's odd I have to call that method and then access an undocumented property to use it...
