### 2022-06-05 pysal/spopt#241 @apulverizer
Consider documenting units for the `max_coverage` parameter of MCLP/LSCP.from_geodataframe
Noticed this as part of https://github.com/openjournals/joss-reviews/issues/3330

It's a bit unclear what the distance unit actually is. I assume it's in the units of the coordinate system that the geodataframes are in. From a usability perspective, it'd be nice to at least document that. Even better, would be to support specifying the units and internally converting it.

Something like:

```python
mclp_from_geodataframe = MCLP.from_geodataframe(
    clients_snapped,
    facilities_snapped,
    "geometry",
    "geometry",
    "demand",
    100,
    service_radius_units="miles",
    p_facilities=P_FACILITIES,
    distance_metric="euclidean",
)
```
