### 2022-06-05 pysal/spopt#242 @apulverizer
Generally usability issues when working with coverage models
As part of https://github.com/openjournals/joss-reviews/issues/3330 I was checking out some of the methods and noticed `get_percentage` causes an exception to be raised

This is just an enhancement request that I think would make it easier for other developers to use this package. As mentioned in #240 there seems to be an order in which methods need to be called - I think this should be documented if the API is not going to change.

For example, in the sample MCLP notebook, I wanted to view the percentage but it raised this error.

```python
Input In [52], in <cell line: 4>()
      2 mclp_from_cost_matrix.client_facility_array()
      3 plot_results(mclp_from_cost_matrix, facility_points)
----> 4 mclp_from_cost_matrix.get_percentage()

File ~/Documents/Source-Code/spopt/spopt/locate/base.py:94, in CoveragePercentageMixin.get_percentage(self)
     90 def get_percentage(self):
     91     """
     92     Calculate the percentage
     93     """
---> 94     self.percentage = 1 - (self.n_cli_uncov / self.aij.shape[0])

AttributeError: 'MCLP' object has no attribute 'n_cli_uncov'

```

I had to dig through the source code to figure out that you need to call `get_uncovered_clients()` first to set `n_cli_uncov`.

----

Would it be possible to make these methods lazily call each other if necessary. So for example, if I call `get_percentage()` before calling it `get_uncovered_clients()`, the `get_percentage()` method would call `get_uncovered_clients()` for me and cache an internal copy of the results for the next time it's called?
