### 2022-06-05 #3330 @apulverizer
@vissarion I've completed my review. I logged a few enhancement requests in the [spopt repo](https://github.com/pysal/spopt/issues/created_by/apulverizer) but nothing that should hold up this paper.

@xf37 @jGaboardi Overall, I found this library very interesting and would like to contribute more to it if I had more time, nice work!
