### 2022-06-05 pysal/spopt#242 @gegen07
@jGaboardi I think here we can call all these functions (`get_uncovered_clients()`, `facility_client_array()` and `client_facility_array`) in solve method, and document the usage of these methods. What do you think?
