### 2022-06-05 pysal/spopt#242 @jGaboardi
I think that's a nice idea. Maybe have a default keyword in `solve()` that automatically calls these. Some thing like:

```python

def solve(self, solver: pulp.LpSolver, results=True):

    self.problem.solve(solver)
    self.check_status()
    if results:
        # call get_uncovered_clients(), facility_client_array() and client_facility_array
    return self
```

But we should probably have that in the `LocateSolver` class. Maybe?
