### 2022-06-05 pysal/spopt#241 @jGaboardi
@apulverizer This is a nice suggestion, and one we have considered before. However, as with other PySAL submodules, we expect the user to have performed preprocessing and to know the projection/coordinate system. Therefore, the user must specify only distance unit itself. Considering the naming [`max_coverage` was recently changed to `service_radius`](https://github.com/pysal/spopt/commit/7bcb35374885b4812dc7845194fd99ee5f0540b4) in order to more accurately reflect facility location terminology.
