### 2022-06-06 #3330 @vissarion
@apulverizer have you tried the command "@editorialbot generate my checklist" ?
