### 2022-06-06 #3330 @martinfleis
@vissarion I think that @apulverizer already resolved that and has a checklist at https://github.com/openjournals/joss-reviews/issues/3330#issuecomment-1146703174
