### 2022-06-06 #3330 @vissarion
I didn't want to do it manually, but that's OK, I copied it above.

@apulverizer please inform us when you are OK with the three issues you opened (the one seems already resolved).
