### 2022-06-06 #3330 @jGaboardi
> @vissarion I've completed my review. I logged a few enhancement requests in the [spopt repo](https://github.com/pysal/spopt/issues/created_by/apulverizer) but nothing that should hold up this paper.
>
> @xf37 @jGaboardi Overall, I found this library very interesting and would like to contribute more to it if I had more time, nice work!

@apulverizer Thanks for your super quick review. @gegen07 and I will look into resolving pysal/spopt#240 and pysal/spopt#242 as soon as possible. Regarding pysal/spopt#241, I left an explanation on why that should be "leave as is" and closed it out.
