### 2022-06-06 bb7a705
add results parameter to solve method
