+++ spopt/locate/coverage.py
@@ -276,7 +276,7 @@ class LSCP(LocateSolver, BaseOutputMixin):
 
             self.fac2cli.append(array_cli)
 
-    def solve(self, solver: pulp.LpSolver):
+    def solve(self, solver: pulp.LpSolver, results: bool = True):
         """
         Solve the LSCP model
 
@@ -285,12 +285,21 @@ class LSCP(LocateSolver, BaseOutputMixin):
         solver: pulp.LpSolver
             solver supported by pulp package
 
+        results: bool
+            if True it will create metainfo - which facilities cover which demand and vice-versa, and the uncovered demand - about the model results
+
         Returns
         -------
         LSCP object
         """
         self.problem.solve(solver)
         self.check_status()
+
+        if results:
+            self.facility_client_array()
+            self.client_facility_array()
+            self.uncovered_clients()
+
         return self
 
 
@@ -605,7 +614,7 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
 
             self.fac2cli.append(array_cli)
 
-    def solve(self, solver: pulp.LpSolver):
+    def solve(self, solver: pulp.LpSolver, results: bool = True):
         """
         Solve the MCLP model
 
@@ -614,10 +623,18 @@ class MCLP(LocateSolver, BaseOutputMixin, CoveragePercentageMixin):
         solver: pulp.LpSolver
             solver supported by pulp package
 
+        results: bool
+            if True it will create metainfo - which facilities cover which demand and vice-versa, and the uncovered demand - about the model results
+
         Returns
         -------
         MCLP object
         """
         self.problem.solve(solver)
         self.check_status()
+
+        if results:
+            self.facility_client_array()
+            self.client_facility_array()
+            self.uncovered_clients()
         return self
