+++ spopt/locate/p_median.py
@@ -316,7 +316,7 @@ class PMedian(LocateSolver, BaseOutputMixin, MeanDistanceMixin):
 
             self.fac2cli.append(array_cli)
 
-    def solve(self, solver: pulp.LpSolver):
+    def solve(self, solver: pulp.LpSolver, results: bool=True):
         """
         Solve the PMedian model
 
@@ -325,10 +325,18 @@ class PMedian(LocateSolver, BaseOutputMixin, MeanDistanceMixin):
         solver: pulp.LpSolver
             solver supported by pulp package
 
+        results: bool
+            if True it will create metainfo - which facilities cover which demand and vice-versa, and the uncovered demand - about the model results
+
         Returns
         -------
         PMedian object
         """
         self.problem.solve(solver)
         self.check_status()
+
+        if results:
+            self.facility_client_array()
+            self.client_facility_array()
+            self.uncovered_clients()
         return self
