### 2022-06-08 93b3672
doc: document class attributes
