+++ spopt/locate/p_median.py
@@ -27,6 +27,19 @@ class PMedian(LocateSolver, BaseOutputMixin, MeanDistanceMixin):
     aij: np.array
         two-dimensional array product of service load/population demand and distance matrix between facility and demand.
 
+    Attributes
+    ----------
+    name: str
+        Problem name
+    problem: pulp.LpProblem
+        Pulp instance of optimization model that contains constraints, variables and objective function.
+    fac2cli : np.array
+        2-d array MxN, where m is number of facilities and n is number of clients. Each row represents a facility and has an array containing clients index meaning that the facility-i cover the entire array.
+    cli2fac: np.array
+        2-d MxN, where m is number of clients and n is number of facilities. Each row represent a client and has an array containing facility index meaning that the client is covered by the facility ith.
+    aij: np.array
+        Cost matrix 2-d array 
+
     """
 
     def __init__(self, name: str, problem: pulp.LpProblem, aij: np.array):
@@ -338,5 +351,5 @@ class PMedian(LocateSolver, BaseOutputMixin, MeanDistanceMixin):
         if results:
             self.facility_client_array()
             self.client_facility_array()
-            self.uncovered_clients()
+
         return self
