### 2022-06-09 ed5688b
chore: add test for attribute class errors
