### 2022-06-10 pysal/spopt#240 @jGaboardi
Resolved via #243.
