### 2022-06-10 1e71809
add spopt ref to JOSS paper
