### 2022-06-10 #3330 @vissarion
@jGaboardi is v0.1.1 the current version? Also could you please archive the package on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other to get a DOI and post it in this issue.
