### 2022-06-10 #3330 @jGaboardi
@vissarion OK, I will see about getting to these edits ASAP and I will touch base again when I have another completed draft. Regarding the version, we are currently on `v0.4.0` and should bump to either `v0.4.1` or `v0.5.0` (@gegen07 do you minor or patch release is appropriate?) to incorporate the changes following @apulverizer's review.
