### 2022-06-11 #3330 @gegen07
@jGaboardi I think it should be okay to bump the version up to `v0.4.1` because I just fixed some errors and documented the class properties better.
