### 2022-06-11 pysal/spopt#246 @jGaboardi
JOSS manuscript clean up
As per @vissarion's [comment](https://github.com/openjournals/joss-reviews/issues/3330#issuecomment-1152117766) in the review:
* [x] add missing DOIs
   * 10.52324/001c.8285 --  PySAL: A Python Library of Spatial Analytical Methods (2007)
   * 10.1111/j.1538-4632.1970.tb00142.x -- Central Facilities Location (1970)
* [x] make hyperlinks clickable
   * decide which can be footnotes...
* [x] reference software repos where possible
