+++ paper/paper.bib
@@ -98,6 +98,7 @@
   number={1},
   pages={5-27},
   keywords={Open Source; Software; Spatial},
+  doi={10.52324/001c.8285},
   url={https://rrs.scholasticahq.com/article/8285.pdf}
 }
 
@@ -157,7 +158,8 @@
   pages={30--42},
   title={{Central Facilities Location}},
   volume={2},
-  year={1970}
+  year={1970},
+  doi={10.1111/j.1538-4632.1970.tb00142.x}
 }
 
 @article{Hakimi1964,
