+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{assunccao2006efficient,
   title={Efficient Regionalization Techniques for Socio-economic Geographical Units Using Minimum Spanning Trees},
-  author={Assun{\c{c}}{\~a}o, Renato M and Neves, Marcos Corr{\^e}a and C{\^a}mara, Gilberto and da Costa Freitas, Corina},
+  author={Assun{\c{c}}{\~a}o, Renato M. and Neves, Marcos Corr{\^e}a and C{\^a}mara, Gilberto and da Costa Freitas, Corina},
   journal={International Journal of Geographical Information Science},
   volume={20},
   number={7},
@@ -26,7 +26,7 @@
 
 @article{koopmans1949optimum,
   title={Optimum utilization of the transportation system},
-  author={Koopmans, Tjalling C},
+  author={Koopmans, Tjalling C.},
   journal={Econometrica: Journal of the Econometric Society},
   pages={136--146},
   year={1949},
@@ -36,7 +36,7 @@
 
 @article{miller1960integer,
   title={Integer programming formulation of traveling salesman problems},
-  author={Miller, Clair E and Tucker, Albert W and Zemlin, Richard A},
+  author={Miller, Clair E. and Tucker, Albert W. and Zemlin, Richard A.},
   journal={Journal of the ACM (JACM)},
   volume={7},
   number={4},
@@ -59,7 +59,7 @@
 
 @article{wei2020efficient,
   title={Efficient Regionalization for Spatially Explicit Neighborhood Delineation},
-  author={Wei, Ran and Rey, Sergio and Knaap, Elijah},
+  author={Wei, Ran and Rey, Sergio J. and Knaap, Elijah},
   journal={International Journal of Geographical Information Science},
   pages={1--17},
   year={2020},
@@ -90,7 +90,7 @@
 }
 
 @article{pysal2007,
-  author={Sergio Rey and Luc Anselin},
+  author={Sergio J. Rey and Luc Anselin},
   title={{PySAL: A Python Library of Spatial Analytical Methods}},
   journal={The Review of Regional Studies},
   year=2007,
@@ -104,7 +104,7 @@
 
 @article{rey2015open,
   title={Open geospatial analytics with PySAL},
-  author={Rey, Sergio J and Anselin, Luc and Li, Xun and Pahle, Robert and Laura, Jason and Li, Wenwen and Koschinsky, Julia},
+  author={Rey, Sergio J. and Anselin, Luc and Li, Xun and Pahle, Robert and Laura, Jason and Li, Wenwen and Koschinsky, Julia},
   journal={ISPRS International Journal of Geo-Information},
   volume={4},
   number={2},
@@ -125,13 +125,13 @@
 
 @article{wolf2020,
   title={Spatially-encouraged spectral clustering: a technique for blending map typologies and regionalization},
-  author={Wolf, Levi},
+  author={Wolf, Levi J.},
   url= {https://osf.io/yzt2p/},
   doi={10.31219/osf.io/yzt2p}
 }
 
 @article{Toregas1971,
-  author={Toregas, Constantine and Swain, R. and ReVelle, C. S. and Bergman, L.},
+  author={Toregas, Constantine and Swain, R.W. and ReVelle, C.S. and Bergman, L.},
   doi={10.1287/opre.19.6.1363},
   journal={Operations Research},
   number={6},
@@ -152,7 +152,7 @@
 }
 
 @article{ReVelle1970,
-  author={ReVelle, C. S. and Swain, R.W.},
+  author={ReVelle, C.S. and Swain, R.W.},
   journal={Geographical Analysis},
   number={1},
   pages={30--42},
@@ -183,3 +183,130 @@
     doi       = {10.5281/zenodo.4444156},
[…]
