### 2022-06-12 pysal/spopt#248 @jGaboardi
JOSS manuscript clean up [2]
xref #246, openjournals/joss-reviews#3330

* [x] clean duplicated URLs in references
* [x] figure 2 caption URL
* [x] code block for Figure 2
   * format for reduced width, line length coming off page
   * possibly reduce figure size?
* [x] clean invalid DOIs
