+++ paper/paper.bib
@@ -201,7 +201,7 @@
   pages={949--958},
   year={2015},
   publisher={Springer},
-  doi={https://doi.org/10.1007/s12145-015-0221-7}
+  doi={10.1007/s12145-015-0221-7}
 }
 
 @article{LEMUSCANOVAS2019114,
@@ -211,7 +211,7 @@
     pages = {114--119},
     year = {2019},
     issn = {1364-8152},
-    doi = {https://doi.org/10.1016/j.envsoft.2019.04.006},
+    doi = {10.1016/j.envsoft.2019.04.006},
     url = {https://www.sciencedirect.com/science/article/pii/S1364815219300064},
     author = {Marc Lemus-Canovas and Joan A. Lopez-Bustins and Javier Martin-Vide and Dominic Royé},
     keywords = {Alps, Environmental regionalization, R package, synoptReg, Synoptic classification},
@@ -257,14 +257,13 @@
     year      = {2009},
     title     = {{nsRFA: Non-supervised Regional Frequency Analysis}},
     url       = {http://www.CRAN.R-project.org/package=nsRFA},
-    howpublished       = {\url{http://www.CRAN.R-project.org/package=nsRFA}},
 }
 
-@Manual{rgeoda,
+@misc{rgeoda,
   title = {rgeoda: R Library for Spatial Data Analysis},
-  author = {Xun Li and Luc Anselin},
+  author = {Li, Xun and Anselin, Luc},
   year = {2022},
-  note = {\url{https://github.com/geodacenter/rgeoda/}, \url{https://geodacenter.github.io/rgeoda/}},
+  url  = {https://github.com/geodacenter/rgeoda/},
 }
 
 @article{rahman2018sdpt3r,
@@ -283,7 +282,6 @@
     year      = {2022},
     title     = {{RBGL: An interface to the BOOST graph library}},
     url       = {https://www.bioconductor.org/packages/release/bioc/html/RBGL.html},
-    howpublished       = {\url{https://www.bioconductor.org/packages/release/bioc/html/RBGL.html}},
 }
 
 @misc{pyspatialopt,
@@ -291,7 +289,6 @@
     year      = {2016},
     title     = {{pyspatialopt: A library to create, solve, and analyze spatial optimization problems}},
     url       = {https://github.com/apulverizer/pyspatialopt},
-    howpublished       = {\url{https://github.com/apulverizer/pyspatialopt}},
 }
 
 @misc{allagash,
@@ -299,7 +296,6 @@
     year      = {2019},
     title     = {{allagash: A spatial optmization library for covering problems}},
     url       = {https://github.com/apulverizer/allagash},
-    howpublished       = {\url{https://github.com/apulverizer/allagash}},
 }
 
 @misc{maxcovr,
@@ -307,6 +303,5 @@
     year      = {2019},
     title     = {{maxcovr: A Set of Tools for Solving the Maximal Covering Location Problem}},
     url       = {https://github.com/njtierney/maxcovr},
-    howpublished       = {\url{https://github.com/njtierney/maxcovr}},
 }
 
