### 2022-06-12 pysal/spopt#248 @jGaboardi
I think the code block in Figure 2 is too large are cumbersome (see current rendering [here](https://github.com/openjournals/joss-papers/blob/joss.03330/joss.03330/10.21105.joss.03330.pdf)). Moreover, it can't be run by "copy&paste"ing the code and creates half-blank page situation in the typesetting. I vote that we remove that code block and point readers to the notebook, which we are already doing in the caption. What are your thoughts @gegen07?
