### 2022-06-12 pysal/spopt#248 @gegen07
TL;DR: I think it's ok to remove the mentioned code block.

When I read a paper, code blocks are, generally, hard to understand and it eventually makes me disinterested. As the tutorial aims at the code explanation and how API works, I think it's ok to remove the code block.
