### 2022-06-12 pysal/spopt#248 @jGaboardi
@gegen07, I also think we could add the following sending to the facility location section:

> Moreover, we have included functionality to place pre-determined site locations within the facility selection pool, allowing for realistic scenarios whereby 1 or more *new* facilities can be added to an *existing* set of locations.

What do you think?
