### 2022-06-12 pysal/spopt#248 @gegen07
That's awesome. I think we can also explicitly add that this is a feature which no of the open source tools provide. What do you think?
