### 2022-06-13 #3330 @arfon
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/2488959009?check_suite_focus=true).

@tarleb – it looks like these failures were due to an issue pulling the container image from Docker. I _had_ thought that we were pulling images from the GitHub container registry these days?
