### 2022-06-13 #3330 @jGaboardi
@vissarion I believe we have complete all requested manuscript clean up. We have:
* cleaned up & added appropriate references
* cleaned up hyperlinks and added some as footnotes
* made several minor addition/edits in the facility location section of the manuscript
* removes to large code block in facility location section
* reviewed for typos & grammar, etc.

Below is the information for the version release associated with the paper:

| title | spopt: a python package for solving spatial optimization problems in PySAL |
|  --- | --- |
| version | 0.4.1 |
| release | https://github.com/pysal/spopt/releases/tag/v0.4.1 |
|  doi | https://doi.org/10.5281/zenodo.6638721 |
