### 2022-06-14 #3330 @jGaboardi
@vissarion

* Zenodo title is now exactly the same
* The links in the footnotes work for me. Do they not work for you?

https://user-images.githubusercontent.com/8590583/173579219-aa1ba185-8ad3-474f-80c5-0da1141ac5c2.mov
