### 2022-06-14 #3330 @vissarion
@jGaboardi hyperlinks do not work for me, unfortunately, in Ubuntu 12 with neither `okular` nor `document viewer` nor `Firefox` but `Chrome` opens them! Anyway maybe it is something local on my side or something with the markdown to pdf conversion.
