### 2022-06-14 #3330 @jGaboardi
I can try to add in the links again and regenerate. If I do this though, will I have to bump to `v0.4.2` and create a new release, etc.?
