### 2022-06-14 #3330 @tarleb
@arfon: The GitHub Container Registry requires users to authenticate before pulling -- that would require some more changes to the action which I haven't implemented yet.

@jGaboardi: Can you try to enclose the links with `<`/`>` as in `<https://example.com>`?
