### 2022-06-14 #3330 @jGaboardi
@vissarion OK, I regenerated the paper and can confirm that the link now open within Acrobat. I had only used Chrome before, but once I checked in Acrobat I can see that the previous versions' footnote links did not work. Should I cut a new `spopt` release and Zenodo Archive?
