### 2022-06-14 #3330 @martinfleis
> Should I cut a new spopt release and Zenodo Archive?

That shouldn't be necessary, I assume. There was no change to the package itself, only to the paper.
